==> 08_figure08_biaoxingzhengli.R <==
rm(list = ls())
download.file("http://togows.dbcls.jp/entry/pathway/hsa04668/genes.json", "data/genesets/hsa04668.json")
download.file("http://togows.dbcls.jp/entry/pathway/hsa04064/genes.json", "data/genesets/hsa04064.json")
download.file("http://togows.dbcls.jp/entry/pathway/hsa04310/genes.json", "data/genesets/hsa04310.json")
download.file("http://togows.dbcls.jp/entry/pathway/hsa04151/genes.json", "data/genesets/hsa04151.json")
download.file("http://togows.dbcls.jp/entry/pathway/hsa04630/genes.json", "data/genesets/hsa04630.json")
download.file("http://togows.dbcls.jp/entry/pathway/hsa04350/genes.json", "data/genesets/hsa04350.json")
download.file("http://togows.dbcls.jp/entry/pathway/hsa04330/genes.json", "data/genesets/hsa04330.json")
download.file("http://togows.dbcls.jp/entry/pathway/hsa04340/genes.json", "data/genesets/hsa04340.json")

download.file("http://togows.dbcls.jp/entry/pathway/hsa04915/genes.json", "data/genesets/hsa04915.json")
download.file("http://togows.dbcls.jp/entry/pathway/hsa03320/genes.json", "data/genesets/hsa03320.json")
download.file("http://togows.dbcls.jp/entry/pathway/hsa04012/genes.json", "data/genesets/hsa04012.json")
download.file("http://togows.dbcls.jp/entry/pathway/hsa04370/genes.json", "data/genesets/hsa04370.json")
download.file("http://togows.dbcls.jp/entry/pathway/hsa04722/genes.json", "data/genesets/hsa04722.json")
download.file("http://togows.dbcls.jp/entry/pathway/hsa04550/genes.json", "data/genesets/hsa04550.json")
download.file("http://togows.dbcls.jp/entry/pathway/hsa04620/genes.json", "data/genesets/hsa04620.json")
download.file("http://togows.dbcls.jp/entry/pathway/hsa04917/genes.json", "data/genesets/hsa04917.json")


download.file("http://togows.dbcls.jp/entry/pathway/hsa04926/genes.json", "data/genesets/hsa04926.json")
download.file("http://togows.dbcls.jp/entry/pathway/hsa04072/genes.json", "data/genesets/hsa04072.json")
download.file("http://togows.dbcls.jp/entry/pathway/hsa04933/genes.json", "data/genesets/hsa04933.json")
download.file("http://togows.dbcls.jp/entry/pathway/hsa04071/genes.json", "data/genesets/hsa04071.json")
download.file("http://togows.dbcls.jp/entry/pathway/hsa04330/genes.json", "data/genesets/hsa04330.json")
download.file("http://togows.dbcls.jp/entry/pathway/hsa04919/genes.json", "data/genesets/hsa04919.json")
download.file("http://togows.dbcls.jp/entry/pathway/hsa04014/genes.json", "data/genesets/hsa04014.json")
download.file("http://togows.dbcls.jp/entry/pathway/hsa04024/genes.json", "data/genesets/hsa04024.json")

download.file("http://togows.dbcls.jp/entry/pathway/hsa04392/genes.json", "data/genesets/hsa04392.json")
download.file("http://togows.dbcls.jp/entry/pathway/hsa04912/genes.json", "data/genesets/hsa04912.json")
download.file("http://togows.dbcls.jp/entry/pathway/hsa04622/genes.json", "data/genesets/hsa04622.json")
download.file("http://togows.dbcls.jp/entry/pathway/hsa04910/genes.json", "data/genesets/hsa04910.json")
download.file("http://togows.dbcls.jp/entry/pathway/hsa04020/genes.json", "data/genesets/hsa04020.json")
download.file("http://togows.dbcls.jp/entry/pathway/hsa04062/genes.json", "data/genesets/hsa04062.json")
download.file("http://togows.dbcls.jp/entry/pathway/hsa04920/genes.json", "data/genesets/hsa04920.json")
download.file("http://togows.dbcls.jp/entry/pathway/hsa04115/genes.json", "data/genesets/hsa04115.json")

download.file("http://togows.dbcls.jp/entry/pathway/hsa04015/genes.json", "data/genesets/hsa04015.json")
download.file("http://togows.dbcls.jp/entry/pathway/hsa04066/genes.json", "data/genesets/hsa04066.json")
download.file("http://togows.dbcls.jp/entry/pathway/hsa04068/genes.json", "data/genesets/hsa04068.json")
download.file("http://togows.dbcls.jp/entry/pathway/hsa04371/genes.json", "data/genesets/hsa04371.json")
download.file("http://togows.dbcls.jp/entry/pathway/hsa04662/genes.json", "data/genesets/hsa04662.json")
download.file("http://togows.dbcls.jp/entry/pathway/hsa04922/genes.json", "data/genesets/hsa04922.json")
download.file("http://togows.dbcls.jp/entry/pathway/hsa04022/genes.json", "data/genesets/hsa04022.json")
download.file("http://togows.dbcls.jp/entry/pathway/hsa04621/genes.json", "data/genesets/hsa04621.json")


download.file("http://togows.dbcls.jp/entry/pathway/hsa04010/genes.json", "data/genesets/hsa04010.json")
download.file("http://togows.dbcls.jp/entry/pathway/hsa04390/genes.json", "data/genesets/hsa04390.json")
download.file("http://togows.dbcls.jp/entry/pathway/hsa04921/genes.json", "data/genesets/hsa04921.json")
download.file("http://togows.dbcls.jp/entry/pathway/hsa04664/genes.json", "data/genesets/hsa04664.json")
download.file("http://togows.dbcls.jp/entry/pathway/hsa04657/genes.json", "data/genesets/hsa04657.json")
download.file("http://togows.dbcls.jp/entry/pathway/hsa04152/genes.json", "data/genesets/hsa04152.json")
download.file("http://togows.dbcls.jp/entry/pathway/hsa04660/genes.json", "data/genesets/hsa04660.json")
download.file("http://togows.dbcls.jp/entry/pathway/hsa04625/genes.json", "data/genesets/hsa04625.json")
download.file("http://togows.dbcls.jp/entry/pathway/hsa04150/genes.json", "data/genesets/hsa04150.json")


json1 = fromJSON(file ="data/genesets/hsa04668.json")
json2 = fromJSON(file ="data/genesets/hsa04064.json")
json3 = fromJSON(file ="data/genesets/hsa04310.json")
json4 = fromJSON(file ="data/genesets/hsa04151.json")
json5 = fromJSON(file ="data/genesets/hsa04630.json")
json6 = fromJSON(file ="data/genesets/hsa04350.json")
json7 = fromJSON(file ="data/genesets/hsa04330.json")
json8 = fromJSON(file ="data/genesets/hsa04340.json")

json9 = fromJSON(file ="data/genesets/hsa04915.json")
json10 = fromJSON(file ="data/genesets/hsa03320.json")
json11 = fromJSON(file ="data/genesets/hsa04012.json")
json12 = fromJSON(file ="data/genesets/hsa04370.json")
json13 = fromJSON(file ="data/genesets/hsa04722.json")
json14 = fromJSON(file ="data/genesets/hsa04550.json")
json15 = fromJSON(file ="data/genesets/hsa04620.json")
json16 = fromJSON(file ="data/genesets/hsa04917.json")

json17= fromJSON(file ="data/genesets/hsa04926.json")
json18 = fromJSON(file ="data/genesets/hsa04072.json")
json19 = fromJSON(file ="data/genesets/hsa04933.json")
json20 = fromJSON(file ="data/genesets/hsa04071.json")
json22 = fromJSON(file ="data/genesets/hsa04919.json")
json23 = fromJSON(file ="data/genesets/hsa04014.json")
json24 = fromJSON(file ="data/genesets/hsa04024.json")

json25 = fromJSON(file ="data/genesets/hsa04392.json")
json26 = fromJSON(file ="data/genesets/hsa04912.json")
json27 = fromJSON(file ="data/genesets/hsa04622.json")
json28 = fromJSON(file ="data/genesets/hsa04910.json")
json29 = fromJSON(file ="data/genesets/hsa04020.json")
json30 = fromJSON(file ="data/genesets/hsa04062.json")
json31 = fromJSON(file ="data/genesets/hsa04920.json")
json32 = fromJSON(file ="data/genesets/hsa04115.json")

json33 = fromJSON(file ="data/genesets/hsa04015.json")
json34 = fromJSON(file ="data/genesets/hsa04066.json")
json35 = fromJSON(file ="data/genesets/hsa04068.json")
json36 = fromJSON(file ="data/genesets/hsa04371.json")
json37 = fromJSON(file ="data/genesets/hsa04662.json")
json38 = fromJSON(file ="data/genesets/hsa04922.json")
json39 = fromJSON(file ="data/genesets/hsa04022.json")
json40 = fromJSON(file ="data/genesets/hsa04621.json")

json41 = fromJSON(file ="data/genesets/hsa04010.json")
json42 = fromJSON(file ="data/genesets/hsa04390.json")
json43 = fromJSON(file ="data/genesets/hsa04921.json")
json44 = fromJSON(file ="data/genesets/hsa04664.json")
json45 = fromJSON(file ="data/genesets/hsa04657.json")
json46 = fromJSON(file ="data/genesets/hsa04152.json")
json47 = fromJSON(file ="data/genesets/hsa04660.json")
json48 = fromJSON(file ="data/genesets/hsa04625.json")
json49 = fromJSON(file ="data/genesets/hsa04150.json")

geneSet1 = list(as.character(sapply(json1[[1]], function(x) sapply(strsplit(x[1], ";"), function(x) x[1]))))[[1]]
geneSet2 = list(as.character(sapply(json2[[1]], function(x) sapply(strsplit(x[1], ";"), function(x) x[1]))))[[1]]
geneSet3 = list(as.character(sapply(json3[[1]], function(x) sapply(strsplit(x[1], ";"), function(x) x[1]))))[[1]]
geneSet4 = list(as.character(sapply(json4[[1]], function(x) sapply(strsplit(x[1], ";"), function(x) x[1]))))[[1]]
geneSet5 = list(as.character(sapply(json5[[1]], function(x) sapply(strsplit(x[1], ";"), function(x) x[1]))))[[1]]
geneSet6 = list(as.character(sapply(json6[[1]], function(x) sapply(strsplit(x[1], ";"), function(x) x[1]))))[[1]]
geneSet7 = list(as.character(sapply(json7[[1]], function(x) sapply(strsplit(x[1], ";"), function(x) x[1]))))[[1]]
geneSet8 = list(as.character(sapply(json8[[1]], function(x) sapply(strsplit(x[1], ";"), function(x) x[1]))))[[1]]

geneSet9 = list(as.character(sapply(json9[[1]], function(x) sapply(strsplit(x[1], ";"), function(x) x[1]))))[[1]]
geneSet10 = list(as.character(sapply(json10[[1]], function(x) sapply(strsplit(x[1], ";"), function(x) x[1]))))[[1]]
geneSet11 = list(as.character(sapply(json11[[1]], function(x) sapply(strsplit(x[1], ";"), function(x) x[1]))))[[1]]
geneSet12 = list(as.character(sapply(json12[[1]], function(x) sapply(strsplit(x[1], ";"), function(x) x[1]))))[[1]]
geneSet13 = list(as.character(sapply(json13[[1]], function(x) sapply(strsplit(x[1], ";"), function(x) x[1]))))[[1]]
geneSet14 = list(as.character(sapply(json14[[1]], function(x) sapply(strsplit(x[1], ";"), function(x) x[1]))))[[1]]
geneSet15 = list(as.character(sapply(json15[[1]], function(x) sapply(strsplit(x[1], ";"), function(x) x[1]))))[[1]]
geneSet16 = list(as.character(sapply(json16[[1]], function(x) sapply(strsplit(x[1], ";"), function(x) x[1]))))[[1]]

geneSet17 = list(as.character(sapply(json17[[1]], function(x) sapply(strsplit(x[1], ";"), function(x) x[1]))))[[1]]
geneSet18 = list(as.character(sapply(json18[[1]], function(x) sapply(strsplit(x[1], ";"), function(x) x[1]))))[[1]]
geneSet19 = list(as.character(sapply(json19[[1]], function(x) sapply(strsplit(x[1], ";"), function(x) x[1]))))[[1]]
geneSet20 = list(as.character(sapply(json20[[1]], function(x) sapply(strsplit(x[1], ";"), function(x) x[1]))))[[1]]
geneSet22 = list(as.character(sapply(json22[[1]], function(x) sapply(strsplit(x[1], ";"), function(x) x[1]))))[[1]]
geneSet23 = list(as.character(sapply(json23[[1]], function(x) sapply(strsplit(x[1], ";"), function(x) x[1]))))[[1]]
geneSet24 = list(as.character(sapply(json24[[1]], function(x) sapply(strsplit(x[1], ";"), function(x) x[1]))))[[1]]

geneSet25 = list(as.character(sapply(json25[[1]], function(x) sapply(strsplit(x[1], ";"), function(x) x[1]))))[[1]]
geneSet26 = list(as.character(sapply(json26[[1]], function(x) sapply(strsplit(x[1], ";"), function(x) x[1]))))[[1]]
geneSet27 = list(as.character(sapply(json27[[1]], function(x) sapply(strsplit(x[1], ";"), function(x) x[1]))))[[1]]
geneSet28 = list(as.character(sapply(json28[[1]], function(x) sapply(strsplit(x[1], ";"), function(x) x[1]))))[[1]]
geneSet29 = list(as.character(sapply(json29[[1]], function(x) sapply(strsplit(x[1], ";"), function(x) x[1]))))[[1]]
geneSet30 = list(as.character(sapply(json30[[1]], function(x) sapply(strsplit(x[1], ";"), function(x) x[1]))))[[1]]
geneSet31 = list(as.character(sapply(json31[[1]], function(x) sapply(strsplit(x[1], ";"), function(x) x[1]))))[[1]]
geneSet32 = list(as.character(sapply(json32[[1]], function(x) sapply(strsplit(x[1], ";"), function(x) x[1]))))[[1]]

geneSet33 = list(as.character(sapply(json33[[1]], function(x) sapply(strsplit(x[1], ";"), function(x) x[1]))))[[1]]
geneSet34 = list(as.character(sapply(json34[[1]], function(x) sapply(strsplit(x[1], ";"), function(x) x[1]))))[[1]]
geneSet35 = list(as.character(sapply(json35[[1]], function(x) sapply(strsplit(x[1], ";"), function(x) x[1]))))[[1]]
geneSet36 = list(as.character(sapply(json36[[1]], function(x) sapply(strsplit(x[1], ";"), function(x) x[1]))))[[1]]
geneSet37 = list(as.character(sapply(json37[[1]], function(x) sapply(strsplit(x[1], ";"), function(x) x[1]))))[[1]]
geneSet38 = list(as.character(sapply(json38[[1]], function(x) sapply(strsplit(x[1], ";"), function(x) x[1]))))[[1]]
geneSet39 = list(as.character(sapply(json39[[1]], function(x) sapply(strsplit(x[1], ";"), function(x) x[1]))))[[1]]
geneSet40 = list(as.character(sapply(json40[[1]], function(x) sapply(strsplit(x[1], ";"), function(x) x[1]))))[[1]]

geneSet41 = list(as.character(sapply(json41[[1]], function(x) sapply(strsplit(x[1], ";"), function(x) x[1]))))[[1]]
geneSet42 = list(as.character(sapply(json42[[1]], function(x) sapply(strsplit(x[1], ";"), function(x) x[1]))))[[1]]
geneSet43 = list(as.character(sapply(json43[[1]], function(x) sapply(strsplit(x[1], ";"), function(x) x[1]))))[[1]]
geneSet44 = list(as.character(sapply(json44[[1]], function(x) sapply(strsplit(x[1], ";"), function(x) x[1]))))[[1]]
geneSet45 = list(as.character(sapply(json45[[1]], function(x) sapply(strsplit(x[1], ";"), function(x) x[1]))))[[1]]
geneSet46 = list(as.character(sapply(json46[[1]], function(x) sapply(strsplit(x[1], ";"), function(x) x[1]))))[[1]]
geneSet47 = list(as.character(sapply(json47[[1]], function(x) sapply(strsplit(x[1], ";"), function(x) x[1]))))[[1]]
geneSet48 = list(as.character(sapply(json48[[1]], function(x) sapply(strsplit(x[1], ";"), function(x) x[1]))))[[1]]
geneSet49 = list(as.character(sapply(json49[[1]], function(x) sapply(strsplit(x[1], ";"), function(x) x[1]))))[[1]]


a = c('TNF signaling pathway','NF-κB signaling pathway','Wnt signaling pathway','PI3K/AKT signaling pathway',
      'JAK-STAT signaling pathway','TGF-β signaling pathway','Notch signaling pathway','Hedgehog signaling pathway',
      'Estrogen signaling pathway','PPAR signaling pathway','ErbB signaling pathway','VEGF signaling pathway',
      'Neurotrophin signaling pathway','Signaling pathways regulating pluripotency of stem cells','Toll-like receptor signaling pathway','Prolactin signaling pathway',
      'Relaxin signaling pathway','Phospholipase D signaling pathway','AGE-RAGE signaling pathway in diabetic complications','Sphingolipid signaling pathway',
      'Thyroid hormone signaling pathway','Ras signaling pathway','cAMP signaling pathway',
      'Hippo signaling pathway - multiple species','GnRH signaling pathway','RIG-I-like receptor signaling pathway','Insulin signaling pathway',
      'Calcium signaling pathway','Chemokine signaling pathway','Adipocytokine signaling pathway','p53 signaling pathway',
      'Rap1 signaling pathway','HIF-1 signaling pathway','FoxO signaling pathway','Apelin signaling pathway','B cell receptor signaling pathway',
      'Glucagon signaling pathway','cGMP-PKG signaling pathway','NOD-like receptor signaling pathway','MAPK signaling pathway',
      'Hippo signaling pathway','Oxytocin signaling pathway','Fc epsilon RI signaling pathway',
      'IL-17 signaling pathway','AMPK signaling pathway','T cell receptor signaling pathway','C-type lectin receptor signaling pathway',
      'mTOR signaling pathway')
pathways <- data.frame()
geneset <- list (geneSet1,geneSet2,geneSet3,geneSet4,geneSet5,geneSet6,geneSet7,geneSet8,geneSet9,geneSet10,
                 geneSet11,geneSet12,geneSet13,geneSet14,geneSet15,geneSet16,geneSet17,geneSet18,geneSet19,geneSet20,
                 geneSet22,geneSet23,geneSet24,geneSet25,geneSet26,geneSet27,geneSet28,geneSet29,geneSet30,
                 geneSet31,geneSet32,geneSet33,geneSet34,geneSet35,geneSet36,geneSet37,geneSet38,geneSet39,geneSet40,
                 geneSet41,geneSet42,geneSet43,geneSet44,geneSet45,geneSet46,geneSet47,geneSet48,geneSet49)

for (i in 1:length(a)) {
        genes =geneset[[i]]
        value = a [i]
        current_pathways = data.frame(value = rep(value, length(genes)), 
                                      genes = genes)
        pathways <- rbind(pathways, current_pathways)
}
length(table(pathways$value))
write.csv(pathways,file = 'figure08_zhengligeneset/pathways_genesets.csv',quote = F,row.names = F)
pathways = split(pathways$genes,pathways$value)
lapply(pathways[1:3], head)
save(pathways,file = 'figure08_zhengligeneset/pathways_genesets.Rdata')

rm(list = ls())
#整理表型
#程序性细胞死亡
pcd <- clusterProfiler::read.gmt('data/phenotype/程序性细胞死亡/程序性细胞死亡genesets.v2023.2.Hs.gmt')
pcd2 <- clusterProfiler::read.gmt('data/phenotype/程序性细胞死亡/程序性细胞死亡genesets.v2023.2.Hs_c2.gmt')
# pcd4 <- clusterProfiler::read.gmt('data/phenotype/程序性细胞死亡/程序性细胞死亡genesets.v2023.2.Hs_c4.gmt')
# pcd5 <- clusterProfiler::read.gmt('data/phenotype/程序性细胞死亡/程序性细胞死亡genesets.v2023.2.Hs_c5.gmt')
# pcd6 <- clusterProfiler::read.gmt('data/phenotype/程序性细胞死亡/程序性细胞死亡genesets.v2023.2.Hs_c6.gmt')
# pcd7 <- clusterProfiler::read.gmt('data/phenotype/程序性细胞死亡/程序性细胞死亡genesets.v2023.2.Hs_c7.gmt')
# pcd8 <- clusterProfiler::read.gmt('data/phenotype/程序性细胞死亡/程序性细胞死亡genesets.v2023.2.Hs_c8.gmt')
#pcd9 <- read.csv('data/phenotype/程序性细胞死亡/程序性细胞死亡GeneCards-SearchResults.csv', header = T,check.names = F)
#table(pcd9$Category)
#pcd9 <- pcd9[pcd9$Category=='Protein Coding',]
pcd10 <- c(pcd$gene,pcd2$gene)
#pcd <- intersect(pcd9$`Gene Symbol`,pcd10)

pcd <- data.frame(Name = rep('Programmed Cell Death',length(pcd10)),genes = pcd10)
save(pcd,file = 'data/phenotype/程序性细胞死亡/pcd_gene.Rdata')
write.csv(pcd, file = 'data/phenotype/程序性细胞死亡/pcd_gene.csv',row.names = F,quote = F)
rm(list = ls())
#非程序性坏死
necro <- clusterProfiler::read.gmt('data/phenotype/非程序性坏死/非程序性坏死genesets.v2023.2.Hs.gmt')
#necro1 <- read.csv('data/phenotype/非程序性坏死/非程序性坏死GeneCards-SearchResults.csv',header = T,check.names = F)
# necro1 <- necro1[necro1$Category=='Protein Coding'&necro1$`Relevance score`>3,]
# necro <- intersect(necro1$`Gene Symbol`,necro$gene)
necro <- data.frame(Name = rep('Necrosis',length(necro$gene)),genes = necro$gene)
save(necro,file = 'data/phenotype/非程序性坏死/necro_gene.Rdata')
write.csv(necro,file = 'data/phenotype/非程序性坏死/necro_gene.csv',row.names = F,quote = F)

rm(list = ls())
#凋亡
apop <- clusterProfiler::read.gmt('data/phenotype/细胞凋亡/细胞凋亡genesets.v2023.2.Hs.gmt')
#apop1 <- read.csv('data/phenotype/细胞凋亡/细胞凋亡GeneCards-SearchResults.csv',header = T,check.names = F)
# table(apop1$Category)
# apop1 <- apop1[apop1$`Relevance score`>3,]
# apop <- intersect(apop$gene,apop1$`Gene Symbol`)
apop <- data.frame(Name = rep('Apoptosis',length(apop$gene)),genes = apop$gene)
save(apop,file = 'data/phenotype/细胞凋亡/apop_gene.Rdata')
write.csv(apop,file = 'data/phenotype/细胞凋亡/apop_gene.csv',row.names = F,quote = F)

rm(list = ls())
#坏死性凋亡
necropto <- clusterProfiler::read.gmt('data/phenotype/坏死性凋亡/坏死性凋亡genesets.v2023.2.Hs.gmt')
#necropto1 <- read.csv('data/phenotype/坏死性凋亡/坏死性凋亡GeneCards-SearchResults.csv',header = T,check.names = F)
# table(necropto1$Category)
# necropto1 <- necropto1[necropto1$`Relevance score`>3,]
# necropto <- intersect(necropto$gene,necropto1$`Gene Symbol`)
necropto <- data.frame(Name = rep('Necroptosis',length(necropto$gene)),genes=necropto$gene)
save(necropto,file = 'data/phenotype/坏死性凋亡/necropto_gene.Rdata')
write.csv(necropto,file = 'data/phenotype/坏死性凋亡/necropto_gene.csv',row.names = F,quote = F)

#细胞焦亡
rm(list = ls())
pyroptosis <- clusterProfiler::read.gmt('data/phenotype/细胞焦亡/细胞焦亡genesets.v2023.2.Hs.gmt')
#pyroptosis1 <- read.csv('data/phenotype/细胞焦亡/细胞焦亡GeneCards-SearchResults.csv',header = T,check.names = F)
# pyroptosis1 <- pyroptosis1[pyroptosis1$`Relevance score`>3,]
# pyroptosis <- intersect(pyroptosis$gene,pyroptosis1$`Gene Symbol`)
pyroptosis <- data.frame(Name = rep('Pyroptosis',length(pyroptosis$gene)),genes = pyroptosis$gene)
save(pyroptosis,file = 'data/phenotype/细胞焦亡/pyroptosis_gene.Rdata')
write.csv(pyroptosis,file = 'data/phenotype/细胞焦亡/pyroptosis_gene.csv',row.names = F,quote = F)

#铁死亡
rm(list = ls())
library(tidyverse)
Ferroptosis <- clusterProfiler::read.gmt('data/phenotype/铁死亡/铁死亡genesets.v2023.2.Hs.gmt')
F1_driver <- read.csv('data/phenotype/铁死亡/ferroptosis_driver.csv',header = T,check.names = F)
F2_driver <- read.csv('data/phenotype/铁死亡/ferroptosis_early_preview_upto20221231/ferroptosis_early_preview_upto20221231/driver.csv')
F_driver = c(F1_driver$symbol, F2_driver$Symbol_or_reported_abbr) 
F_driver = F_driver[!duplicated(F_driver)]
F1_marker <- read.csv('data/phenotype/铁死亡/ferroptosis_marker.csv',header = T,check.names = T)
F2_marker <- read.csv('data/phenotype/铁死亡/ferroptosis_early_preview_upto20221231/ferroptosis_early_preview_upto20221231/marker.csv')
F_marker = c(F1_marker$symbol, F2_marker$Symbol) 
F_marker = F_marker[!duplicated(F_marker)]
F1_sup <- read.csv('data/phenotype/铁死亡/ferroptosis_suppressor.csv',header = T,check.names = T)
F2_sup <- read.csv('data/phenotype/铁死亡/ferroptosis_early_preview_upto20221231/ferroptosis_early_preview_upto20221231/suppressor.csv')
F_sup = c(F1_sup$symbol, F2_sup$Symbol) 
F_sup = F_sup[!duplicated(F_sup)]
# F_genecard <- read.csv(file = 'data/phenotype/铁死亡/铁死亡GeneCards-SearchResults.csv')
# F_genecard <- F_genecard[F_genecard$Relevance.score > 3,]
Ferrd <- c(F_sup,F_marker,F_driver)
Ferro <- c(Ferroptosis$gene,Ferrd)
Ferro <- data.frame(Name = rep('Ferroptosis',length(Ferro)),genes = Ferro)
save(Ferro,file = 'data/phenotype/铁死亡/Ferro.Rdata')
write.csv(Ferro,file = 'data/phenotype/铁死亡/Ferro.csv',row.names = F,quote = F)

#铜死亡
rm(list = ls())
cupp1 <- read.csv(file = 'data/phenotype/铁死亡/铁死亡GeneCards-SearchResults.csv',check.names = F,header = T)
cupp1 <- cupp1[cupp1$`Relevance score`>3,]
cupp <- cupp1$`Gene Symbol`
cupp <- data.frame(Name = rep('Cuproptosis',length(cupp)),genes=cupp)
save(cupp,file = 'data/phenotype/铜死亡/cupp.Rdata')
write.csv(cupp,file = 'data/phenotype/铜死亡/cupp.csv',row.names = F,quote = F)


#双硫死亡
rm(list = ls())
disu <- openxlsx::read.xlsx('data/phenotype/双硫死亡/Disulfidptosisext_20230626.xlsx')
disu <- disu[disu$RCD=='Disulfidptosis',]
disu1 <- openxlsx::read.xlsx('data/phenotype/双硫死亡/Disulfidptosis_literaturePMID 38035071.xlsx')
disu <- c(disu$Symbol_or_reported_abbr,disu1$genesymbol)
disu <- data.frame(Name = rep('Disulfidptosis',length(disu)),genes=disu)
save(disu,file = 'data/phenotype/双硫死亡/disu.Rdata')
write.csv(disu,file = 'data/phenotype/双硫死亡/disu.csv',row.names = F,quote = F)

#免疫原性细胞死亡
rm(list = ls())
Immune <- clusterProfiler::read.gmt('data/phenotype/免疫原性细胞死亡/genesets.v2023.2.Hs.gmt')
Immune2 <- clusterProfiler::read.gmt('data/phenotype/免疫原性细胞死亡/genesets.v2023.2.Hs_c2.gmt')
# Immune4 <- clusterProfiler::read.gmt('data/phenotype/免疫原性细胞死亡/genesets.v2023.2.Hs_c4.gmt')
# Immune5 <- clusterProfiler::read.gmt('data/phenotype/免疫原性细胞死亡/genesets.v2023.2.Hs_c5.gmt')
# Immune6 <- clusterProfiler::read.gmt('data/phenotype/免疫原性细胞死亡/genesets.v2023.2.Hs_c6.gmt')
# Immune7 <- clusterProfiler::read.gmt('data/phenotype/免疫原性细胞死亡/genesets.v2023.2.Hs_c7.gmt')
# Immune8 <- clusterProfiler::read.gmt('data/phenotype/免疫原性细胞死亡/genesets.v2023.2.Hs_c8.gmt')
# Immune9 <- read.csv('data/phenotype/免疫原性细胞死亡/免疫原性的细胞死亡GeneCards-SearchResults.csv', header = T, check.names = F)
# Immune9 <- Immune9[Immune9$`Relevance score`>3,]
immune1 <- c(Immune$gene,Immune2$gene)
immune <- c(immune1)
immune <- data.frame(Name = rep('Immunogenic cell death',length(immune)),genes = immune)
save(immune,file = 'data/phenotype/免疫原性细胞死亡/immune_gene.Rdata')
write.csv(immune,file = 'data/phenotype/免疫原性细胞死亡/immune_gene.csv',row.names = F,quote = F)

#增殖
rm(list = ls())
proliferation = clusterProfiler::read.gmt('data/phenotype/增殖/增殖genesets.v2023.2.Hs.gmt')
proliferation <- data.frame(Name = rep('Proliferation',length(proliferation$gene)),genes = proliferation$gene)
save(proliferation,file = 'data/phenotype/增殖/proliferation_gene.Rdata')
write.csv(proliferation,file = 'data/phenotype/增殖/proliferation_gene.csv',row.names = F,quote = F)

#迁移
rm(list = ls())
migration = clusterProfiler::read.gmt('data/phenotype/迁移/迁移genesets.v2023.2.Hs.gmt')
migration <- data.frame(Name = rep('Migration',length(migration$gene)),genes = migration$gene)
save(migration,file = 'data/phenotype/迁移/migration_gene.Rdata')
write.csv(migration,file = 'data/phenotype/迁移/migration_gene.csv',row.names = F,quote = F)

#侵袭
rm(list = ls())
invasion = clusterProfiler::read.gmt('data/phenotype/侵袭/侵袭genesets.v2023.2.Hs.gmt')
invasion <- data.frame(Name = rep('Invasion',length(invasion$gene)),genes = invasion$gene)
save(invasion,file = 'data/phenotype/侵袭/invasion_gene.Rdata')
write.csv(invasion,file = 'data/phenotype/侵袭/invasion_gene.csv',row.names = F,quote = F)


rm(list = ls())
#糖酵解
gly <- clusterProfiler::read.gmt('data/phenotype/糖酵解/糖酵解genesets.v2023.2.Hs.gmt')
# #gly1 <- read.csv('data/phenotype/糖酵解/糖酵解GeneCards-SearchResults.csv',check.names = F,header = T)
# gly1 <- gly1[gly1$`Relevance score`>3,]
# gly <- intersect(gly$gene,gly1$`Gene Symbol`)
gly <- data.frame(Name = rep('Glycolysis',length(gly$gene)),genes = gly$gene)
save(gly,file = 'data/phenotype/糖酵解/gly_gene.Rdata')
write.csv(gly,file = 'data/phenotype/糖酵解/gly_gene.csv',quote = F,row.names = F)


#脂肪酸代谢
rm(list = ls())
fatty <- clusterProfiler::read.gmt('data/phenotype/脂肪酸代谢/脂肪酸代谢genesets.v2023.2.Hs.gmt')
# fatty1 <- read.csv('data/phenotype/脂肪酸代谢/脂肪酸代谢GeneCards-SearchResults.csv',check.names = F,header = T)
# fatty1 <- fatty1[fatty1$`Relevance score`>3,]
# fatty <- intersect(fatty$gene,fatty1$`Gene Symbol`)
fatty <- data.frame(Name = rep('Fatty acid metabolism',length(fatty$gene)),genes = fatty$gene)
save(fatty,file = 'data/phenotype/脂肪酸代谢/fatty_gene.Rdata')
write.csv(fatty,file = 'data/phenotype/脂肪酸代谢/fatty_gene.csv',quote = F,row.names = F)

#脂质代谢
rm(list = ls())
lip <- clusterProfiler::read.gmt('data/phenotype/脂质代谢/脂质代谢genesets.v2023.2.Hs.gmt')
# lip1 <- read.csv('data/phenotype/脂质代谢/脂质代谢GeneCards-SearchResults.csv',check.names = F,header = T)
# lip1 <- lip1[lip1$`Relevance score`>3,]
# lip <- intersect(lip$gene,lip1$`Gene Symbol`)
lip <- data.frame(Name = rep('Lipid metabolism',length(lip$gene)),genes = lip$gene)
save(lip,file = 'data/phenotype/脂质代谢/lip_gene.Rdata')
write.csv(lip,file = 'data/phenotype/脂质代谢/lip_gene.csv',quote = F,row.names = F)

#胆固醇代谢
rm(list = ls())
cho <- clusterProfiler::read.gmt('data/phenotype/胆固醇代谢/胆固醇代谢genesets.v2023.2.Hs.gmt')
# cho1 <- read.csv('data/phenotype/胆固醇代谢/胆固醇代谢GeneCards-SearchResults.csv',check.names = F,header = T)
# cho1 <- cho1[cho1$`Relevance score`>3,]
# cho <- intersect(cho$gene,cho1$`Gene Symbol`)
cho <- data.frame(Name = rep('Cholesterol metabolism',length(cho$gene)),genes = cho$gene)
save(cho,file = 'data/phenotype/胆固醇代谢/cho_gene.Rdata')
write.csv(cho,file = 'data/phenotype/胆固醇代谢/cho_gene.csv',quote = F,row.names = F)

#氨基酸代谢
rm(list = ls())
aa <- clusterProfiler::read.gmt('data/phenotype/氨基酸代谢/氨基酸代谢genesets.v2023.2.Hs.gmt')
# aa1 <- read.csv('data/phenotype/氨基酸代谢/氨基酸代谢GeneCards-SearchResults.csv',check.names = F,header = T)
# aa1 <- aa1[aa1$`Relevance score`>3,]
# aa <- intersect(aa$gene,aa1$`Gene Symbol`)
aa <- data.frame(Name = rep('Amino acid metabolism',length(aa$gene)),genes = aa$gene)
save(aa,file = 'data/phenotype/氨基酸代谢/aa_gene.Rdata')
write.csv(aa,file = 'data/phenotype/氨基酸代谢/aa_gene.csv',quote = F,row.names = F)


#代谢重编程
rm(list = ls())
me <- clusterProfiler::read.gmt('data/phenotype/代谢重编程/代谢重编程genesets.v2023.2.Hs.gmt')
# me1 <- read.csv('data/phenotype/代谢重编程/代谢重编程GeneCards-SearchResults.csv',check.names = F,header = T)
# me1 <- me1[me1$`Relevance score`>3,]
# me <- intersect(me$gene,me1$`Gene Symbol`)
me <- data.frame(Name = rep('Metabolic Reprogramming',length(me$gene)),genes = me$gene)
save(me,file = 'data/phenotype/代谢重编程/me_gene.Rdata')
write.csv(me,file = 'data/phenotype/代谢重编程/me_gene.csv',quote = F,row.names = F)

#缺氧
rm(list = ls())
ox <- clusterProfiler::read.gmt('data/phenotype/缺氧/缺氧genesets.v2023.2.Hs.gmt')
# ox1 <- read.csv('data/phenotype/缺氧/缺氧GeneCards-SearchResults.csv',check.names = F,header = T)
# ox1 <- ox1[ox1$`Relevance score`>3,]
# ox <- intersect(ox$gene,ox1$`Gene Symbol`)
ox <- data.frame(Name = rep('Hypoxia',length(ox$gene)),genes = ox$gene)
save(ox,file = 'data/phenotype/缺氧/ox_gene.Rdata')
write.csv(ox,file = 'data/phenotype/缺氧/ox_gene.csv',quote = F,row.names = F)

#内质网应激
rm(list = ls())
endo <- clusterProfiler::read.gmt('data/phenotype/内质网应激/内质网应激genesets.v2023.2.Hs.gmt')
# endo1 <- read.csv('data/phenotype/内质网应激/内质网应激GeneCards-SearchResults.csv',check.names = F,header = T)
# endo1 <- endo1[endo1$`Relevance score`>3,]
# endo <- intersect(endo$gene,endo1$`Gene Symbol`)
endo <- data.frame(Name = rep('Endoplasmic Reticulum stress',length(endo$gene)),genes = endo$gene)
save(endo,file = 'data/phenotype/内质网应激/endo_gene.Rdata')
write.csv(endo,file = 'data/phenotype/内质网应激/endo_gene.csv',quote = F,row.names = F)


#氧化应激
rm(list = ls())
oxi <- clusterProfiler::read.gmt('data/phenotype/氧化应激/氧化应激genesets.v2023.2.Hs.gmt')
# oxi1 <- read.csv('data/phenotype/内质网应激/内质网应激GeneCards-SearchResults.csv',check.names = F,header = T)
# oxi1 <- oxi1[oxi1$`Relevance score`>3,]
# oxi <- intersect(oxi$gene,oxi1$`Gene Symbol`)
oxi <- data.frame(Name = rep('Oxidative stress',length(oxi$gene)),genes = oxi$gene)
save(oxi,file = 'data/phenotype/氧化应激/oxi_gene.Rdata')
write.csv(oxi,file = 'data/phenotype/氧化应激/oxi_gene.csv',quote = F,row.names = F)


#衰老
rm(list = ls())
age <- clusterProfiler::read.gmt('data/phenotype/衰老/衰老genesets.v2023.2.Hs.gmt')
# age1 <- read.csv('data/phenotype/衰老/衰老GeneCards-SearchResults.csv',check.names = F,header = T)
# age1 <- age1[age1$`Relevance score`>3,]
# age <- intersect(age$gene,age1$`Gene Symbol`)
age <- data.frame(Name = rep('Aging',length(age$gene)),genes = age$gene)
save(age,file = 'data/phenotype/衰老/age_gene.Rdata')
write.csv(age,file = 'data/phenotype/衰老/age_gene.csv',quote = F,row.names = F)


#EMT
rm(list = ls())
emt <- clusterProfiler::read.gmt('data/phenotype/EMT/EMTgenesets.v2023.2.Hs.gmt')
# emt1 <- read.csv('data/phenotype/EMT/EMTGeneCards-SearchResults.csv',check.names = F,header = T)
# emt1 <- emt1[emt1$`Relevance score`>3,]
# emt <- intersect(emt$gene,emt1$`Gene Symbol`)
emt <- data.frame(Name = rep('Epithelial-mesenchymal transition',length(emt$gene)),genes = emt$gene)
save(emt,file = 'data/phenotype/EMT/emt_gene.Rdata')
write.csv(emt,file = 'data/phenotype/EMT/emt_gene.csv',quote = F,row.names = F)

#昼夜节律
rm(list = ls())
circ <- clusterProfiler::read.gmt('data/phenotype/昼夜节律/昼夜节律genesets.v2023.2.Hs.gmt')
# circ1 <- read.csv('data/phenotype/昼夜节律/昼夜节律GeneCards-SearchResults.csv',check.names = F,header = T)
# circ1 <- circ1[circ1$`Relevance score`>3,]
# circ <- intersect(circ$gene,circ1$`Gene Symbol`)
circ <- data.frame(Name = rep('Circadian rhythms',length(circ$gene)),genes = circ$gene)
save(circ,file = 'data/phenotype/昼夜节律/circ_gene.Rdata')
write.csv(circ,file = 'data/phenotype/昼夜节律/circ_gene.csv',quote = F,row.names = F)

#细胞干性
rm(list = ls())
stem <- clusterProfiler::read.gmt('data/phenotype/细胞干性/细胞干性genesets.v2023.2.Hs.gmt')
# stem1 <- read.csv('data/phenotype/细胞干性/细胞干性GeneCards-SearchResults.csv',check.names = F,header = T)
# stem1 <- stem1[stem1$`Relevance score`>3,]
# stem <- stem1$`Gene Symbol`
stem <- data.frame(Name = rep('Cell stemness',length(stem$gene)),genes = stem$gene)
save(stem,file = 'data/phenotype/细胞干性/stem_gene.Rdata')
write.csv(stem,file = 'data/phenotype/细胞干性/stem_gene.csv',quote = F,row.names = F)

#自噬
rm(list = ls())
auto <- clusterProfiler::read.gmt('data/phenotype/自噬/自噬genesets.v2023.2.Hs.gmt')
# auto1 <- read.csv('data/phenotype/自噬/自噬GeneCards-SearchResults.csv',check.names = F,header = T)
# auto1 <- auto1[auto1$`Relevance score`>3,]
# auto <- intersect(auto1$`Gene Symbol`,auto$gene)
auto <- data.frame(Name = rep('Autophagy',length(auto$gene)),genes = auto$gene)
save(auto,file = 'data/phenotype/自噬/auto_gene.Rdata')
write.csv(auto,file = 'data/phenotype/自噬/auto_gene.csv',quote = F,row.names = F)

#线粒体自噬
rm(list = ls())
mito <- clusterProfiler::read.gmt('data/phenotype/线粒体自噬/线粒体自噬genesets.v2023.2.Hs.gmt')
# mito1 <- read.csv('data/phenotype/线粒体自噬/线粒体自噬GeneCards-SearchResults.csv',check.names = F,header = T)
# mito1 <- mito1[mito1$`Relevance score`>3,]
# mito <- intersect(mito1$`Gene Symbol`,mito$gene)
mito <- data.frame(Name = rep('Mitophagy',length(mito$gene)),genes = mito$gene)
save(mito,file = 'data/phenotype/线粒体自噬/mito_gene.Rdata')
write.csv(mito,file = 'data/phenotype/线粒体自噬/mito_gene.csv',quote = F,row.names = F)


rm(list = ls())
m1a <- openxlsx::read.xlsx('data/phenotype/m1aPMID 38115786.xlsx')
m1a <- data.frame(Name = rep('m1A',length(m1a$gene)),genes = m1a$gene)
m5c <- openxlsx::read.xlsx('data/phenotype/m5cPMID 36911744PMID 38270308.xlsx')
m5c <- data.frame(Name = rep('m5C',length(m5c$gene)),genes = m5c$gene)
m6a <- openxlsx::read.xlsx('data/phenotype/m6APMID 38610981.xlsx')
m6a <- data.frame(Name = rep('m6A',length(m6a$gene)),genes = m6a$gene)
m7g <- openxlsx::read.xlsx('data/phenotype/m7GPMID：36726408.xlsx')
m7g <- data.frame(Name = rep('m7G',length(m7g$gene)),genes = m7g$gene)
mRNA <- rbind(m1a,rbind(m5c,rbind(m6a,m7g)))
save(mRNA,file = 'data/phenotype/mRNA.Rdata')

#死亡形式
load('data/phenotype/程序性细胞死亡/pcd_gene.Rdata')
load('data/phenotype/非程序性坏死/necro_gene.Rdata')
load('data/phenotype/细胞凋亡/apop_gene.Rdata')
load('data/phenotype/坏死性凋亡/necropto_gene.Rdata')
load('data/phenotype/细胞焦亡/pyroptosis_gene.Rdata')
load('data/phenotype/铁死亡/Ferro.Rdata')
load('data/phenotype/铜死亡/cupp.Rdata')
#load('data/phenotype/双硫死亡/disu.Rdata')
load('data/phenotype/免疫原性细胞死亡/immune_gene.Rdata')
df1 <- rbind(pcd,rbind(necro,rbind(apop,rbind(necropto,rbind(pyroptosis,rbind(Ferro,rbind(cupp,immune)))))))
table(df1$Name)


#自噬
load('data/phenotype/自噬/auto_gene.Rdata')
load('data/phenotype/线粒体自噬/mito_gene.Rdata')
df2 <- rbind(auto,mito)
table(df2$Name)
#代谢
load('data/phenotype/糖酵解/gly_gene.Rdata')
load('data/phenotype/脂肪酸代谢/fatty_gene.Rdata')
load('data/phenotype/脂质代谢/lip_gene.Rdata')
load('data/phenotype/胆固醇代谢/cho_gene.Rdata')
load('data/phenotype/氨基酸代谢/aa_gene.Rdata')
load('data/phenotype/代谢重编程/me_gene.Rdata')
df3 <- rbind(gly,rbind(fatty,rbind(lip,rbind(cho,rbind(aa,me)))))
table(df3$Name)
  
#其他
load('data/phenotype/缺氧/ox_gene.Rdata')
load('data/phenotype/内质网应激/endo_gene.Rdata')
load('data/phenotype/氧化应激/oxi_gene.Rdata')
load('data/phenotype/衰老/age_gene.Rdata')
load('data/phenotype/昼夜节律/circ_gene.Rdata')
load('data/phenotype/细胞干性/stem_gene.Rdata')
df4 <- rbind(ox,rbind(endo,rbind(oxi,rbind(age,rbind(circ,stem)))))
table(df4$Name)

#增殖相关
#load('data/phenotype/增殖/proliferation_gene.Rdata')
load('data/phenotype/迁移/migration_gene.Rdata')
load('data/phenotype/侵袭/invasion_gene.Rdata')
load('data/phenotype/EMT/emt_gene.Rdata')
df5 <- rbind(emt,rbind(invasion,migration))
table(df5)


df <- rbind(df1,rbind(df2,rbind(df3,rbind(df4,rbind(df5,mRNA)))))

write.csv(df,file = 'figure08_zhengligeneset/phenotype_genesets.csv',row.names = F,quote = F)
df <- split(df$genes,df$Name)
head(df,3)
save(df,file = 'figure08_zhengligeneset/phenotype_genesets.Rdata')

###########1.1计算tcga表型estimate评分ssgsea####
rm(list = ls())
library(GSVA)
load('figure08_zhengligeneset/phenotype_genesets.Rdata')
load('figure02_E2F6/lihc_sur_data.Rdata')
exprSet[1:4,1:4]
load('figure06_two_group/tcga_predict_factor_clinicalinformation.Rdata')
identical(colnames(exprSet),rownames(meta))
exp=exprSet
exp[1:4,1:4]

geneset <- df
class(exp)
exp=as.matrix(exp)####这个一定要是矩阵，不能是数据框
f = "figure08_zhengligeneset/tcga_phenotype_ssgsea.Rdata"
if(!file.exists(f)){
  re <- gsva(exp, geneset, method="ssgsea",
             mx.diff=FALSE, verbose=FALSE)
  save(re,file = f)
}
load(f)
library(pheatmap)
Cluster1_name=rownames(meta)[meta$Cluster=='Cluster1']
Cluster2_name=rownames(meta)[meta$Cluster=='Cluster2']
table(meta$Group)
Group1_name=rownames(meta)[meta$Group=='Others']
Group2_name=rownames(meta)[meta$Group=='E2F6_H&TOP2A_H']
re1=re[,Cluster1_name]
re2=re[,Cluster2_name]
re=cbind(re1,re2)
meta=meta[match(colnames(re),rownames(meta)),]
identical(rownames(meta),colnames(re))
re2=as.data.frame(re)
an = data.frame(group = meta$Cluster,
                row.names = colnames(re2))
dev.off()
re2=as.matrix(re2)
bk <- c(seq(-2,-0.1,by=0.01),seq(0,2,by=0.01))
pdf('figure08_zhengligeneset/tcga_ssgsea_phenotype_heatmap_cluster.pdf',height = 5,width = 8)
pheatmap(re2,scale = "row",
         show_colnames = F,
         annotation_col = an,
         cluster_cols = F,
         cluster_rows = F,column_split=an,
         #legend_breaks=seq(-2,2,1),
         breaks=bk,
         color = colorRampPalette(c("navy", "white", "firebrick3"))(100))

dev.off()

identical(colnames(re),rownames(meta))
Group=meta$Cluster
draw_boxplot(re,Group,xlab = ' ',ylab = 'Score',sort = F)
ggsave('figure08_zhengligeneset/tcga_ssgsea_phenotype_boxplot_cluster.pdf',height = 6,width = 8,units = 'in')

Group1_name=rownames(meta)[meta$Group=='Others']
Group2_name=rownames(meta)[meta$Group=='E2F6_H&TOP2A_H']
re1=re[,Group1_name]
re2=re[,Group2_name]
re=cbind(re1,re2)
meta=meta[match(colnames(re),rownames(meta)),]
identical(rownames(meta),colnames(re))
re2=as.data.frame(re)
an = data.frame(group = meta$Group,
                row.names = colnames(re2))
dev.off()
re2=as.matrix(re2)
bk <- c(seq(-2,-0.1,by=0.01),seq(0,2,by=0.01))
pdf('figure08_zhengligeneset/tcga_ssgsea_phenotype_heatmap.pdf',height = 5,width = 9)
pheatmap(re2,scale = "row",
         show_colnames = F,
         annotation_col = an,
         cluster_cols = F,
         cluster_rows = F,column_split=an,
         #legend_breaks=seq(-2,2,1),
         breaks=bk,
         color = colorRampPalette(c("navy", "white", "firebrick3"))(100))

dev.off()

identical(colnames(re),rownames(meta))
Group=meta$Group
draw_boxplot(re,Group,xlab = ' ',ylab = 'Score',sort = F)
ggsave('figure08_zhengligeneset/tcga_ssgsea_phenotype_boxplot.pdf',height = 6,width = 8,units = 'in')

identical(rownames(meta),colnames(re))
exp=exp[,match(colnames(re),colnames(exp))]
identical(colnames(re),colnames(exp))
exp=exp[c('E2F6','TOP2A'),]
identical(colnames(re),colnames(exp))
re=rbind(re,exp)
identical(colnames(exp),rownames(meta))
identical(colnames(re),rownames(meta))
data=as.data.frame(t(re))
data$PredictScore=meta$Predict_score
M=data

library(tidyverse)
library(corrplot)
if(!require(ggcor))devtools::install_github("xukaili/ggcor",upgrade = F)
datp <- cor.mtest(M)$p[1:29,30:32] %>%
  as.data.frame()%>% 
  rownames_to_column("A") %>% 
  gather("B","p",-A)
data <- cor(M)[1:29,30:32] %>%
  as.data.frame()%>% 
  rownames_to_column("A") %>% 
  gather("B","r",-A) %>% 
  mutate(p = datp$p)
data$ps = case_when(data$p<0.01~"**",
                    data$p<0.05~"*",
                    T~"")
library(vctrs)
library(grid)

geom_rectriangle <- function(mapping = NULL, data = NULL,
                             stat = "identity", position = "identity",
                             ...,
                             linejoin = "mitre",
                             na.rm = FALSE,
                             show.legend = NA,
                             inherit.aes = TRUE) {
  layer(
    data = data,
    mapping = mapping,
    stat = stat,
    geom = GeomRectriangle,
    position = position,
    show.legend = show.legend,
    inherit.aes = inherit.aes,
    params = list(
      linejoin = linejoin,
      na.rm = na.rm,
      ...
    )
  )
}

GeomRectriangle <- ggproto(
  "GeomRectriangle", Geom,
  default_aes = aes(r = 1, colour = "grey35", fill = NA, size = 0.25, linetype = 1,
                    alpha = NA,type = "upper"),
  required_aes = c("x", "y"),
  draw_panel = function(self, data, panel_params, coord, linejoin = "mitre",type = "upper") {
    aesthetics <- setdiff(names(data), c("x", "y"))
    
    polys <- lapply(split(data, seq_len(nrow(data))), function(row) {
      rectriangle <- point_to_rectriangle(row$x, row$y, row$r, row$type)
      aes <- new_data_frame(row[aesthetics])[rep(1, 4), ]
      GeomPolygon$draw_panel(cbind(rectriangle, aes), panel_params, coord)
    })
    
    ggplot2:::ggname("geom_rectriangle", do.call("grobTree", polys))
  },
  draw_key = draw_key_polygon
)


point_to_rectriangle <- function(x, y, r, type = type) {
  r <- 0.5 * sign(r) * sqrt(abs(r))
  #r0 = 0.5
  xmin <- - r + x
  xmax <- r + x
  ymin <- - r + y
  ymax <- r + y 
  if(type == "upper"){
    df = new_data_frame(list(
      y = c(ymax, ymax, ymin, ymax),
      x = c(xmin, xmax, xmin, xmin)
    ))
  }else if(type == "lower"){
    df = new_data_frame(list(
      y = c(ymax, ymin, ymin, ymax),
      x = c(xmax, xmax, xmin, xmax) 
    ))
  }
  df
}

data$B=factor(data$B,levels = c('E2F6','TOP2A','PredictScore'))
table(data$B)
ggplot()+
  geom_rectriangle(data = data, aes(A, B, fill = -log10(p)),type = "upper", r = 1)+
  scale_fill_gradient(low = "white", high = "#a5d3ed",na.value = "white")+
  ggnewscale::new_scale_fill()+
  geom_rectriangle(data = data, aes(A, B, fill = r),type = "lower", r = 1)+
  scale_fill_gradient2(high = "#e7857b", mid = "white",low = "#6d9cc5")+
  labs(x = "", y = "")+
  geom_text(data = data,aes(A,B,label = ps),
            nudge_y = 0.05)+
  theme_bw()+
  theme(axis.text.x = element_text(angle = 80, hjust = 1, vjust = 1))
ggsave('figure08_zhengligeneset/tcga_ssgsea_correlation_hubgene_predict_phenotype.pdf',height = 4,width =8,dpi = 320,units = 'in')

##########1.2计算tcga通路estimate评分ssgsea####
rm(list = ls())
library(GSVA)
load('figure08_zhengligeneset/pathways_genesets.Rdata')
load('figure02_E2F6/lihc_sur_data.Rdata')
exprSet[1:4,1:4]
load('figure06_two_group/tcga_predict_factor_clinicalinformation.Rdata')
identical(colnames(exprSet),rownames(meta))
exp=exprSet
exp[1:4,1:4]
geneset <- pathways
class(exp)
exp=as.matrix(exp)####这个一定要是矩阵，不能是数据框
f = "figure08_zhengligeneset/tcga_pathway_ssgsea.Rdata"
if(!file.exists(f)){
  re <- gsva(exp, geneset, method="ssgsea",
             mx.diff=FALSE, verbose=FALSE)
  save(re,file = f)
}
load(f)
library(pheatmap)
Cluster1_name=rownames(meta)[meta$Cluster=='Cluster1']
Cluster2_name=rownames(meta)[meta$Cluster=='Cluster2']
table(meta$Group)
Group1_name=rownames(meta)[meta$Group=='Others']
Group2_name=rownames(meta)[meta$Group=='E2F6_H&TOP2A_H']
re1=re[,Cluster1_name]
re2=re[,Cluster2_name]
re=cbind(re1,re2)
meta=meta[match(colnames(re),rownames(meta)),]
identical(rownames(meta),colnames(re))
re2=as.data.frame(re)
an = data.frame(group = meta$Cluster,
                row.names = colnames(re2))
dev.off()
re2=as.matrix(re2)
bk <- c(seq(-2,-0.1,by=0.01),seq(0,2,by=0.01))
pdf('figure08_zhengligeneset/tcga_ssgsea_pathway_heatmap_cluster.pdf',height = 8,width = 8)
pheatmap(re2,scale = "row",
         show_colnames = F,
         annotation_col = an,
         cluster_cols = F,
         cluster_rows = F,column_split=an,
         #legend_breaks=seq(-2,2,1),
         breaks=bk,
         color = colorRampPalette(c("navy", "white", "firebrick3"))(100))

dev.off()

identical(colnames(re),rownames(meta))
Group=meta$Cluster
draw_boxplot(re,Group,xlab = ' ',ylab = 'Score',sort = F)
ggsave('figure08_zhengligeneset/tcga_ssgsea_pathway_boxplot_cluster.pdf',height = 9,width = 10,units = 'in')

Group1_name=rownames(meta)[meta$Group=='Others']
Group2_name=rownames(meta)[meta$Group=='E2F6_H&TOP2A_H']
re1=re[,Group1_name]
re2=re[,Group2_name]
re=cbind(re1,re2)
meta=meta[match(colnames(re),rownames(meta)),]
identical(rownames(meta),colnames(re))
re2=as.data.frame(re)
an = data.frame(group = meta$Group,
                row.names = colnames(re2))
dev.off()
re2=as.matrix(re2)
bk <- c(seq(-2,-0.1,by=0.01),seq(0,2,by=0.01))
pdf('figure08_zhengligeneset/tcga_ssgsea_pathway_heatmap.pdf',height = 8,width = 10)
pheatmap(re2,scale = "row",
         show_colnames = F,
         annotation_col = an,
         cluster_cols = F,
         cluster_rows = F,column_split=an,
         #legend_breaks=seq(-2,2,1),
         breaks=bk,
         color = colorRampPalette(c("navy", "white", "firebrick3"))(100))

dev.off()

identical(colnames(re),rownames(meta))
Group=meta$Group
draw_boxplot(re,Group,xlab = ' ',ylab = 'Score',sort = F)
ggsave('figure08_zhengligeneset/tcga_ssgsea_pathway_boxplot.pdf',height = 9,width = 12,units = 'in')

identical(rownames(meta),colnames(re))
exp=exp[,match(colnames(re),colnames(exp))]
identical(colnames(re),colnames(exp))
exp=exp[c('E2F6','TOP2A'),]
identical(colnames(re),colnames(exp))
re=rbind(re,exp)
identical(colnames(exp),rownames(meta))
identical(colnames(re),rownames(meta))
data=as.data.frame(t(re))
data$PredictScore=meta$Predict_score
M=data

library(tidyverse)
library(corrplot)
if(!require(ggcor))devtools::install_github("xukaili/ggcor",upgrade = F)
datp <- cor.mtest(M)$p[1:48,49:51] %>%
  as.data.frame()%>% 
  rownames_to_column("A") %>% 
  gather("B","p",-A)
data <- cor(M)[1:48,49:51] %>%
  as.data.frame()%>% 
  rownames_to_column("A") %>% 
  gather("B","r",-A) %>% 
  mutate(p = datp$p)
data$ps = case_when(data$p<0.01~"**",
                    data$p<0.05~"*",
                    T~"")
library(vctrs)
library(grid)

geom_rectriangle <- function(mapping = NULL, data = NULL,
                             stat = "identity", position = "identity",
                             ...,
                             linejoin = "mitre",
                             na.rm = FALSE,
                             show.legend = NA,
                             inherit.aes = TRUE) {
  layer(
    data = data,
    mapping = mapping,
    stat = stat,
    geom = GeomRectriangle,
    position = position,
    show.legend = show.legend,
    inherit.aes = inherit.aes,
    params = list(
      linejoin = linejoin,
      na.rm = na.rm,
      ...
    )
  )
}

GeomRectriangle <- ggproto(
  "GeomRectriangle", Geom,
  default_aes = aes(r = 1, colour = "grey35", fill = NA, size = 0.25, linetype = 1,
                    alpha = NA,type = "upper"),
  required_aes = c("x", "y"),
  draw_panel = function(self, data, panel_params, coord, linejoin = "mitre",type = "upper") {
    aesthetics <- setdiff(names(data), c("x", "y"))
    
    polys <- lapply(split(data, seq_len(nrow(data))), function(row) {
      rectriangle <- point_to_rectriangle(row$x, row$y, row$r, row$type)
      aes <- new_data_frame(row[aesthetics])[rep(1, 4), ]
      GeomPolygon$draw_panel(cbind(rectriangle, aes), panel_params, coord)
    })
    
    ggplot2:::ggname("geom_rectriangle", do.call("grobTree", polys))
  },
  draw_key = draw_key_polygon
)


point_to_rectriangle <- function(x, y, r, type = type) {
  r <- 0.5 * sign(r) * sqrt(abs(r))
  #r0 = 0.5
  xmin <- - r + x
  xmax <- r + x
  ymin <- - r + y
  ymax <- r + y 
  if(type == "upper"){
    df = new_data_frame(list(
      y = c(ymax, ymax, ymin, ymax),
      x = c(xmin, xmax, xmin, xmin)
    ))
  }else if(type == "lower"){
    df = new_data_frame(list(
      y = c(ymax, ymin, ymin, ymax),
      x = c(xmax, xmax, xmin, xmax) 
    ))
  }
  df
}

data$B=factor(data$B,levels = c('E2F6','TOP2A','PredictScore'))
table(data$B)
ggplot()+
  geom_rectriangle(data = data, aes(A, B, fill = -log10(p)),type = "upper", r = 1)+
  scale_fill_gradient(low = "white", high = "#a5d3ed",na.value = "white")+
  ggnewscale::new_scale_fill()+
  geom_rectriangle(data = data, aes(A, B, fill = r),type = "lower", r = 1)+
  scale_fill_gradient2(high = "#e7857b", mid = "white",low = "#6d9cc5")+
  labs(x = "", y = "")+
  geom_text(data = data,aes(A,B,label = ps),
            nudge_y = 0.05)+
  theme_bw()+
  theme(axis.text.x = element_text(angle = 80, hjust = 1, vjust = 1))
ggsave('figure08_zhengligeneset/tcga_ssgsea_correlation_hubgene_predict_pathway.pdf',height = 6,width =12,dpi = 320,units = 'in')

###########1.3计算tcga表型estimate评分gsva####
rm(list = ls())
library(GSVA)
load('figure08_zhengligeneset/phenotype_genesets.Rdata')
load('figure02_E2F6/lihc_sur_data.Rdata')
exprSet[1:4,1:4]
load('figure06_two_group/tcga_predict_factor_clinicalinformation.Rdata')
identical(colnames(exprSet),rownames(meta))
exp=exprSet
exp[1:4,1:4]
geneset <- df
class(exp)
exp=as.matrix(exp)####这个一定要是矩阵，不能是数据框
f = "figure08_zhengligeneset/tcga_phenotype_gsva.Rdata"
if(!file.exists(f)){
  re <- gsva(exp, geneset, method="gsva",
             mx.diff=FALSE, verbose=FALSE)
  save(re,file = f)
}
load(f)
library(pheatmap)
Cluster1_name=rownames(meta)[meta$Cluster=='Cluster1']
Cluster2_name=rownames(meta)[meta$Cluster=='Cluster2']
table(meta$Group)
Group1_name=rownames(meta)[meta$Group=='Others']
Group2_name=rownames(meta)[meta$Group=='E2F6_H&TOP2A_H']
re1=re[,Cluster1_name]
re2=re[,Cluster2_name]
re=cbind(re1,re2)
meta=meta[match(colnames(re),rownames(meta)),]
identical(rownames(meta),colnames(re))
re2=as.data.frame(re)
an = data.frame(group = meta$Cluster,
                row.names = colnames(re2))
dev.off()
re2=as.matrix(re2)
bk <- c(seq(-2,-0.1,by=0.01),seq(0,2,by=0.01))
pdf('figure08_zhengligeneset/tcga_gsva_phenotype_heatmap_cluster.pdf',height = 5,width = 8)
pheatmap(re2,scale = "row",
         show_colnames = F,
         annotation_col = an,
         cluster_cols = F,
         cluster_rows = F,column_split=an,
         #legend_breaks=seq(-2,2,1),
         breaks=bk,
         color = colorRampPalette(c("navy", "white", "firebrick3"))(100))

dev.off()

identical(colnames(re),rownames(meta))
Group=meta$Cluster
draw_boxplot(re,Group,xlab = ' ',ylab = 'Score',sort = F)
ggsave('figure08_zhengligeneset/tcga_gsva_phenotype_boxplot_cluster.pdf',height = 6,width = 8,units = 'in')

Group1_name=rownames(meta)[meta$Group=='Others']
Group2_name=rownames(meta)[meta$Group=='E2F6_H&TOP2A_H']
re1=re[,Group1_name]
re2=re[,Group2_name]
re=cbind(re1,re2)
meta=meta[match(colnames(re),rownames(meta)),]
identical(rownames(meta),colnames(re))
re2=as.data.frame(re)
an = data.frame(group = meta$Group,
                row.names = colnames(re2))
dev.off()
re2=as.matrix(re2)
bk <- c(seq(-2,-0.1,by=0.01),seq(0,2,by=0.01))
pdf('figure08_zhengligeneset/tcga_gsva_phenotype_heatmap.pdf',height = 5,width = 9)
pheatmap(re2,scale = "row",
         show_colnames = F,
         annotation_col = an,
         cluster_cols = F,
         cluster_rows = F,column_split=an,
         #legend_breaks=seq(-2,2,1),
         breaks=bk,
         color = colorRampPalette(c("navy", "white", "firebrick3"))(100))

dev.off()

identical(colnames(re),rownames(meta))
Group=meta$Group
draw_boxplot(re,Group,xlab = ' ',ylab = 'Score',sort = F)
ggsave('figure08_zhengligeneset/tcga_gsva_phenotype_boxplot.pdf',height = 6,width = 8,units = 'in')

identical(rownames(meta),colnames(re))
exp=exp[,match(colnames(re),colnames(exp))]
identical(colnames(re),colnames(exp))
exp=exp[c('E2F6','TOP2A'),]
identical(colnames(re),colnames(exp))
re=rbind(re,exp)
identical(colnames(exp),rownames(meta))
identical(colnames(re),rownames(meta))
data=as.data.frame(t(re))
data$PredictScore=meta$Predict_score
M=data

library(tidyverse)
library(corrplot)
if(!require(ggcor))devtools::install_github("xukaili/ggcor",upgrade = F)
datp <- cor.mtest(M)$p[1:29,30:32] %>%
  as.data.frame()%>% 
  rownames_to_column("A") %>% 
  gather("B","p",-A)
data <- cor(M)[1:29,30:32] %>%
  as.data.frame()%>% 
  rownames_to_column("A") %>% 
  gather("B","r",-A) %>% 
  mutate(p = datp$p)
data$ps = case_when(data$p<0.01~"**",
                    data$p<0.05~"*",
                    T~"")
library(vctrs)
library(grid)

geom_rectriangle <- function(mapping = NULL, data = NULL,
                             stat = "identity", position = "identity",
                             ...,
                             linejoin = "mitre",
                             na.rm = FALSE,
                             show.legend = NA,
                             inherit.aes = TRUE) {
  layer(
    data = data,
    mapping = mapping,
    stat = stat,
    geom = GeomRectriangle,
    position = position,
    show.legend = show.legend,
    inherit.aes = inherit.aes,
    params = list(
      linejoin = linejoin,
      na.rm = na.rm,
      ...
    )
  )
}

GeomRectriangle <- ggproto(
  "GeomRectriangle", Geom,
  default_aes = aes(r = 1, colour = "grey35", fill = NA, size = 0.25, linetype = 1,
                    alpha = NA,type = "upper"),
  required_aes = c("x", "y"),
  draw_panel = function(self, data, panel_params, coord, linejoin = "mitre",type = "upper") {
    aesthetics <- setdiff(names(data), c("x", "y"))
    
    polys <- lapply(split(data, seq_len(nrow(data))), function(row) {
      rectriangle <- point_to_rectriangle(row$x, row$y, row$r, row$type)
      aes <- new_data_frame(row[aesthetics])[rep(1, 4), ]
      GeomPolygon$draw_panel(cbind(rectriangle, aes), panel_params, coord)
    })
    
    ggplot2:::ggname("geom_rectriangle", do.call("grobTree", polys))
  },
  draw_key = draw_key_polygon
)


point_to_rectriangle <- function(x, y, r, type = type) {
  r <- 0.5 * sign(r) * sqrt(abs(r))
  #r0 = 0.5
  xmin <- - r + x
  xmax <- r + x
  ymin <- - r + y
  ymax <- r + y 
  if(type == "upper"){
    df = new_data_frame(list(
      y = c(ymax, ymax, ymin, ymax),
      x = c(xmin, xmax, xmin, xmin)
    ))
  }else if(type == "lower"){
    df = new_data_frame(list(
      y = c(ymax, ymin, ymin, ymax),
      x = c(xmax, xmax, xmin, xmax) 
    ))
  }
  df
}

data$B=factor(data$B,levels = c('E2F6','TOP2A','PredictScore'))
table(data$B)
ggplot()+
  geom_rectriangle(data = data, aes(A, B, fill = -log10(p)),type = "upper", r = 1)+
  scale_fill_gradient(low = "white", high = "#a5d3ed",na.value = "white")+
  ggnewscale::new_scale_fill()+
  geom_rectriangle(data = data, aes(A, B, fill = r),type = "lower", r = 1)+
  scale_fill_gradient2(high = "#e7857b", mid = "white",low = "#6d9cc5")+
  labs(x = "", y = "")+
  geom_text(data = data,aes(A,B,label = ps),
            nudge_y = 0.05)+
  theme_bw()+
  theme(axis.text.x = element_text(angle = 80, hjust = 1, vjust = 1))
ggsave('figure08_zhengligeneset/tcga_gsva_correlation_hubgene_predict_phenotype.pdf',height = 4,width =8,dpi = 320,units = 'in')

##########1.4计算tcga通路estimate评分gsva####
rm(list = ls())
library(GSVA)
load('figure08_zhengligeneset/pathways_genesets.Rdata')
load('figure02_E2F6/lihc_sur_data.Rdata')
exprSet[1:4,1:4]
load('figure06_two_group/tcga_predict_factor_clinicalinformation.Rdata')
identical(colnames(exprSet),rownames(meta))
exp=exprSet
exp[1:4,1:4]
geneset <- pathways
class(exp)
exp=as.matrix(exp)####这个一定要是矩阵，不能是数据框
f = "figure08_zhengligeneset/tcga_pathway_gsva.Rdata"
if(!file.exists(f)){
  re <- gsva(exp, geneset, method="gsva",
             mx.diff=FALSE, verbose=FALSE)
  save(re,file = f)
}
load(f)
library(pheatmap)
Cluster1_name=rownames(meta)[meta$Cluster=='Cluster1']
Cluster2_name=rownames(meta)[meta$Cluster=='Cluster2']
table(meta$Group)
Group1_name=rownames(meta)[meta$Group=='Others']
Group2_name=rownames(meta)[meta$Group=='E2F6_H&TOP2A_H']
re1=re[,Cluster1_name]
re2=re[,Cluster2_name]
re=cbind(re1,re2)
meta=meta[match(colnames(re),rownames(meta)),]
identical(rownames(meta),colnames(re))
re2=as.data.frame(re)
an = data.frame(group = meta$Cluster,
                row.names = colnames(re2))
dev.off()
re2=as.matrix(re2)
bk <- c(seq(-2,-0.1,by=0.01),seq(0,2,by=0.01))
pdf('figure08_zhengligeneset/tcga_gsva_pathway_heatmap_cluster.pdf',height = 8,width = 8)
pheatmap(re2,scale = "row",
         show_colnames = F,
         annotation_col = an,
         cluster_cols = F,
         cluster_rows = F,column_split=an,
         #legend_breaks=seq(-2,2,1),
         breaks=bk,
         color = colorRampPalette(c("navy", "white", "firebrick3"))(100))

dev.off()

identical(colnames(re),rownames(meta))
Group=meta$Cluster
draw_boxplot(re,Group,xlab = ' ',ylab = 'Score',sort = F)
ggsave('figure08_zhengligeneset/tcga_gsva_pathway_boxplot_cluster.pdf',height = 9,width = 10,units = 'in')

Group1_name=rownames(meta)[meta$Group=='Others']
Group2_name=rownames(meta)[meta$Group=='E2F6_H&TOP2A_H']
re1=re[,Group1_name]
re2=re[,Group2_name]
re=cbind(re1,re2)
meta=meta[match(colnames(re),rownames(meta)),]
identical(rownames(meta),colnames(re))
re2=as.data.frame(re)
an = data.frame(group = meta$Group,
                row.names = colnames(re2))
dev.off()
re2=as.matrix(re2)
bk <- c(seq(-2,-0.1,by=0.01),seq(0,2,by=0.01))
pdf('figure08_zhengligeneset/tcga_gsva_pathway_heatmap.pdf',height = 8,width = 10)
pheatmap(re2,scale = "row",
         show_colnames = F,
         annotation_col = an,
         cluster_cols = F,
         cluster_rows = F,column_split=an,
         #legend_breaks=seq(-2,2,1),
         breaks=bk,
         color = colorRampPalette(c("navy", "white", "firebrick3"))(100))

dev.off()

identical(colnames(re),rownames(meta))
Group=meta$Group
draw_boxplot(re,Group,xlab = ' ',ylab = 'Score',sort = F)
ggsave('figure08_zhengligeneset/tcga_gsva_pathway_boxplot.pdf',height = 9,width = 12,units = 'in')

identical(rownames(meta),colnames(re))
exp=exp[,match(colnames(re),colnames(exp))]
identical(colnames(re),colnames(exp))
exp=exp[c('E2F6','TOP2A'),]
identical(colnames(re),colnames(exp))
re=rbind(re,exp)
identical(colnames(exp),rownames(meta))
identical(colnames(re),rownames(meta))
data=as.data.frame(t(re))
data$PredictScore=meta$Predict_score
M=data

library(tidyverse)
library(corrplot)
if(!require(ggcor))devtools::install_github("xukaili/ggcor",upgrade = F)
datp <- cor.mtest(M)$p[1:48,49:51] %>%
  as.data.frame()%>% 
  rownames_to_column("A") %>% 
  gather("B","p",-A)
data <- cor(M)[1:48,49:51] %>%
  as.data.frame()%>% 
  rownames_to_column("A") %>% 
  gather("B","r",-A) %>% 
  mutate(p = datp$p)
data$ps = case_when(data$p<0.01~"**",
                    data$p<0.05~"*",
                    T~"")
library(vctrs)
library(grid)

geom_rectriangle <- function(mapping = NULL, data = NULL,
                             stat = "identity", position = "identity",
                             ...,
                             linejoin = "mitre",
                             na.rm = FALSE,
                             show.legend = NA,
                             inherit.aes = TRUE) {
  layer(
    data = data,
    mapping = mapping,
    stat = stat,
    geom = GeomRectriangle,
    position = position,
    show.legend = show.legend,
    inherit.aes = inherit.aes,
    params = list(
      linejoin = linejoin,
      na.rm = na.rm,
      ...
    )
  )
}

GeomRectriangle <- ggproto(
  "GeomRectriangle", Geom,
  default_aes = aes(r = 1, colour = "grey35", fill = NA, size = 0.25, linetype = 1,
                    alpha = NA,type = "upper"),
  required_aes = c("x", "y"),
  draw_panel = function(self, data, panel_params, coord, linejoin = "mitre",type = "upper") {
    aesthetics <- setdiff(names(data), c("x", "y"))
    
    polys <- lapply(split(data, seq_len(nrow(data))), function(row) {
      rectriangle <- point_to_rectriangle(row$x, row$y, row$r, row$type)
      aes <- new_data_frame(row[aesthetics])[rep(1, 4), ]
      GeomPolygon$draw_panel(cbind(rectriangle, aes), panel_params, coord)
    })
    
    ggplot2:::ggname("geom_rectriangle", do.call("grobTree", polys))
  },
  draw_key = draw_key_polygon
)


point_to_rectriangle <- function(x, y, r, type = type) {
  r <- 0.5 * sign(r) * sqrt(abs(r))
  #r0 = 0.5
  xmin <- - r + x
  xmax <- r + x
  ymin <- - r + y
  ymax <- r + y 
  if(type == "upper"){
    df = new_data_frame(list(
      y = c(ymax, ymax, ymin, ymax),
      x = c(xmin, xmax, xmin, xmin)
    ))
  }else if(type == "lower"){
    df = new_data_frame(list(
      y = c(ymax, ymin, ymin, ymax),
      x = c(xmax, xmax, xmin, xmax) 
    ))
  }
  df
}

data$B=factor(data$B,levels = c('E2F6','TOP2A','PredictScore'))
table(data$B)
ggplot()+
  geom_rectriangle(data = data, aes(A, B, fill = -log10(p)),type = "upper", r = 1)+
  scale_fill_gradient(low = "white", high = "#a5d3ed",na.value = "white")+
  ggnewscale::new_scale_fill()+
  geom_rectriangle(data = data, aes(A, B, fill = r),type = "lower", r = 1)+
  scale_fill_gradient2(high = "#e7857b", mid = "white",low = "#6d9cc5")+
  labs(x = "", y = "")+
  geom_text(data = data,aes(A,B,label = ps),
            nudge_y = 0.05)+
  theme_bw()+
  theme(axis.text.x = element_text(angle = 80, hjust = 1, vjust = 1))
ggsave('figure08_zhengligeneset/tcga_gsva_correlation_hubgene_predict_pathway.pdf',height = 6,width =12,dpi = 320,units = 'in')

###########2.1计算icgc表型estimate评分ssgsea####
rm(list = ls())
library(GSVA)
load('figure08_zhengligeneset/phenotype_genesets.Rdata')
load('figure02_E2F6/icgc_lihc_sur_data.Rdata')
exprSet[1:4,1:4]
load('figure06_two_group/icgc_predict_factor_clinicalinformation.Rdata')
identical(colnames(exprSet),rownames(meta))
exp=exprSet
exp[1:4,1:4]
geneset <- df
class(exp)
exp=as.matrix(exp)####这个一定要是矩阵，不能是数据框
f = "figure08_zhengligeneset/icgc_phenotype_ssgsea.Rdata"
if(!file.exists(f)){
  re <- gsva(exp, geneset, method="ssgsea",
             mx.diff=FALSE, verbose=FALSE)
  save(re,file = f)
}
load(f)
library(pheatmap)
Cluster1_name=rownames(meta)[meta$Cluster=='Cluster1']
Cluster2_name=rownames(meta)[meta$Cluster=='Cluster2']
table(meta$Group)
Group1_name=rownames(meta)[meta$Group=='Others']
Group2_name=rownames(meta)[meta$Group=='E2F6_H&TOP2A_H']
re1=re[,Cluster1_name]
re2=re[,Cluster2_name]
re=cbind(re1,re2)
meta=meta[match(colnames(re),rownames(meta)),]
identical(rownames(meta),colnames(re))
re2=as.data.frame(re)
an = data.frame(group = meta$Cluster,
                row.names = colnames(re2))
dev.off()
re2=as.matrix(re2)
bk <- c(seq(-2,-0.1,by=0.01),seq(0,2,by=0.01))
pdf('figure08_zhengligeneset/icgc_ssgsea_phenotype_heatmap_cluster.pdf',height = 5,width = 8)
pheatmap(re2,scale = "row",
         show_colnames = F,
         annotation_col = an,
         cluster_cols = F,
         cluster_rows = F,column_split=an,
         #legend_breaks=seq(-2,2,1),
         breaks=bk,
         color = colorRampPalette(c("navy", "white", "firebrick3"))(100))

dev.off()

identical(colnames(re),rownames(meta))
Group=meta$Cluster
draw_boxplot(re,Group,xlab = ' ',ylab = 'Score',sort = F)
ggsave('figure08_zhengligeneset/icgc_ssgsea_phenotype_boxplot_cluster.pdf',height = 6,width = 8,units = 'in')

Group1_name=rownames(meta)[meta$Group=='Others']
Group2_name=rownames(meta)[meta$Group=='E2F6_H&TOP2A_H']
re1=re[,Group1_name]
re2=re[,Group2_name]
re=cbind(re1,re2)
meta=meta[match(colnames(re),rownames(meta)),]
identical(rownames(meta),colnames(re))
re2=as.data.frame(re)
an = data.frame(group = meta$Group,
                row.names = colnames(re2))
dev.off()
re2=as.matrix(re2)
bk <- c(seq(-2,-0.1,by=0.01),seq(0,2,by=0.01))
pdf('figure08_zhengligeneset/icgc_ssgsea_phenotype_heatmap.pdf',height = 5,width = 9)
pheatmap(re2,scale = "row",
         show_colnames = F,
         annotation_col = an,
         cluster_cols = F,
         cluster_rows = F,column_split=an,
         #legend_breaks=seq(-2,2,1),
         breaks=bk,
         color = colorRampPalette(c("navy", "white", "firebrick3"))(100))

dev.off()

identical(colnames(re),rownames(meta))
Group=meta$Group
draw_boxplot(re,Group,xlab = ' ',ylab = 'Score',sort = F)
ggsave('figure08_zhengligeneset/icgc_ssgsea_phenotype_boxplot.pdf',height = 6,width = 8,units = 'in')

identical(rownames(meta),colnames(re))
exp=exp[,match(colnames(re),colnames(exp))]
identical(colnames(re),colnames(exp))
exp=exp[c('E2F6','TOP2A'),]
identical(colnames(re),colnames(exp))
re=rbind(re,exp)
identical(colnames(exp),rownames(meta))
identical(colnames(re),rownames(meta))
data=as.data.frame(t(re))
data$PredictScore=meta$Predict_score
M=data

library(tidyverse)
library(corrplot)
if(!require(ggcor))devtools::install_github("xukaili/ggcor",upgrade = F)
datp <- cor.mtest(M)$p[1:29,30:32] %>%
  as.data.frame()%>% 
  rownames_to_column("A") %>% 
  gather("B","p",-A)
data <- cor(M)[1:29,30:32] %>%
  as.data.frame()%>% 
  rownames_to_column("A") %>% 
  gather("B","r",-A) %>% 
  mutate(p = datp$p)
data$ps = case_when(data$p<0.01~"**",
                    data$p<0.05~"*",
                    T~"")
library(vctrs)
library(grid)

geom_rectriangle <- function(mapping = NULL, data = NULL,
                             stat = "identity", position = "identity",
                             ...,
                             linejoin = "mitre",
                             na.rm = FALSE,
                             show.legend = NA,
                             inherit.aes = TRUE) {
  layer(
    data = data,
    mapping = mapping,
    stat = stat,
    geom = GeomRectriangle,
    position = position,
    show.legend = show.legend,
    inherit.aes = inherit.aes,
    params = list(
      linejoin = linejoin,
      na.rm = na.rm,
      ...
    )
  )
}

GeomRectriangle <- ggproto(
  "GeomRectriangle", Geom,
  default_aes = aes(r = 1, colour = "grey35", fill = NA, size = 0.25, linetype = 1,
                    alpha = NA,type = "upper"),
  required_aes = c("x", "y"),
  draw_panel = function(self, data, panel_params, coord, linejoin = "mitre",type = "upper") {
    aesthetics <- setdiff(names(data), c("x", "y"))
    
    polys <- lapply(split(data, seq_len(nrow(data))), function(row) {
      rectriangle <- point_to_rectriangle(row$x, row$y, row$r, row$type)
      aes <- new_data_frame(row[aesthetics])[rep(1, 4), ]
      GeomPolygon$draw_panel(cbind(rectriangle, aes), panel_params, coord)
    })
    
    ggplot2:::ggname("geom_rectriangle", do.call("grobTree", polys))
  },
  draw_key = draw_key_polygon
)


point_to_rectriangle <- function(x, y, r, type = type) {
  r <- 0.5 * sign(r) * sqrt(abs(r))
  #r0 = 0.5
  xmin <- - r + x
  xmax <- r + x
  ymin <- - r + y
  ymax <- r + y 
  if(type == "upper"){
    df = new_data_frame(list(
      y = c(ymax, ymax, ymin, ymax),
      x = c(xmin, xmax, xmin, xmin)
    ))
  }else if(type == "lower"){
    df = new_data_frame(list(
      y = c(ymax, ymin, ymin, ymax),
      x = c(xmax, xmax, xmin, xmax) 
    ))
  }
  df
}

data$B=factor(data$B,levels = c('E2F6','TOP2A','PredictScore'))
table(data$B)
ggplot()+
  geom_rectriangle(data = data, aes(A, B, fill = -log10(p)),type = "upper", r = 1)+
  scale_fill_gradient(low = "white", high = "#a5d3ed",na.value = "white")+
  ggnewscale::new_scale_fill()+
  geom_rectriangle(data = data, aes(A, B, fill = r),type = "lower", r = 1)+
  scale_fill_gradient2(high = "#e7857b", mid = "white",low = "#6d9cc5")+
  labs(x = "", y = "")+
  geom_text(data = data,aes(A,B,label = ps),
            nudge_y = 0.05)+
  theme_bw()+
  theme(axis.text.x = element_text(angle = 80, hjust = 1, vjust = 1))
ggsave('figure08_zhengligeneset/icgc_ssgsea_correlation_hubgene_predict_phenotype.pdf',height = 4,width =8,dpi = 320,units = 'in')

##########2.2计算icgc通路estimate评分ssgsea####
rm(list = ls())
library(GSVA)
load('figure08_zhengligeneset/pathways_genesets.Rdata')
load('figure02_E2F6/icgc_lihc_sur_data.Rdata')
exprSet[1:4,1:4]
load('figure06_two_group/icgc_predict_factor_clinicalinformation.Rdata')
identical(colnames(exprSet),rownames(meta))
exp=exprSet
exp[1:4,1:4]
geneset <- pathways
class(exp)
exp=as.matrix(exp)####这个一定要是矩阵，不能是数据框
f = "figure08_zhengligeneset/icgc_pathway_ssgsea.Rdata"
if(!file.exists(f)){
  re <- gsva(exp, geneset, method="ssgsea",
             mx.diff=FALSE, verbose=FALSE)
  save(re,file = f)
}
load(f)
library(pheatmap)
Cluster1_name=rownames(meta)[meta$Cluster=='Cluster1']
Cluster2_name=rownames(meta)[meta$Cluster=='Cluster2']
table(meta$Group)
Group1_name=rownames(meta)[meta$Group=='Others']
Group2_name=rownames(meta)[meta$Group=='E2F6_H&TOP2A_H']
re1=re[,Cluster1_name]
re2=re[,Cluster2_name]
re=cbind(re1,re2)
meta=meta[match(colnames(re),rownames(meta)),]
identical(rownames(meta),colnames(re))
re2=as.data.frame(re)
an = data.frame(group = meta$Cluster,
                row.names = colnames(re2))
dev.off()
re2=as.matrix(re2)
bk <- c(seq(-2,-0.1,by=0.01),seq(0,2,by=0.01))
pdf('figure08_zhengligeneset/icgc_ssgsea_pathway_heatmap_cluster.pdf',height = 8,width = 8)
pheatmap(re2,scale = "row",
         show_colnames = F,
         annotation_col = an,
         cluster_cols = F,
         cluster_rows = F,column_split=an,
         #legend_breaks=seq(-2,2,1),
         breaks=bk,
         color = colorRampPalette(c("navy", "white", "firebrick3"))(100))

dev.off()

identical(colnames(re),rownames(meta))
Group=meta$Cluster
draw_boxplot(re,Group,xlab = ' ',ylab = 'Score',sort = F)
ggsave('figure08_zhengligeneset/icgc_ssgsea_pathway_boxplot_cluster.pdf',height = 9,width = 10,units = 'in')

Group1_name=rownames(meta)[meta$Group=='Others']
Group2_name=rownames(meta)[meta$Group=='E2F6_H&TOP2A_H']
re1=re[,Group1_name]
re2=re[,Group2_name]
re=cbind(re1,re2)
meta=meta[match(colnames(re),rownames(meta)),]
identical(rownames(meta),colnames(re))
re2=as.data.frame(re)
an = data.frame(group = meta$Group,
                row.names = colnames(re2))
dev.off()
re2=as.matrix(re2)
bk <- c(seq(-2,-0.1,by=0.01),seq(0,2,by=0.01))
pdf('figure08_zhengligeneset/icgc_ssgsea_pathway_heatmap.pdf',height = 8,width = 10)
pheatmap(re2,scale = "row",
         show_colnames = F,
         annotation_col = an,
         cluster_cols = F,
         cluster_rows = F,column_split=an,
         #legend_breaks=seq(-2,2,1),
         breaks=bk,
         color = colorRampPalette(c("navy", "white", "firebrick3"))(100))

dev.off()

identical(colnames(re),rownames(meta))
Group=meta$Group
draw_boxplot(re,Group,xlab = ' ',ylab = 'Score',sort = F)
ggsave('figure08_zhengligeneset/icgc_ssgsea_pathway_boxplot.pdf',height = 9,width = 12,units = 'in')

identical(rownames(meta),colnames(re))
exp=exp[,match(colnames(re),colnames(exp))]
identical(colnames(re),colnames(exp))
exp=exp[c('E2F6','TOP2A'),]
identical(colnames(re),colnames(exp))
re=rbind(re,exp)
identical(colnames(exp),rownames(meta))
identical(colnames(re),rownames(meta))
data=as.data.frame(t(re))
data$PredictScore=meta$Predict_score
M=data

library(tidyverse)
library(corrplot)
if(!require(ggcor))devtools::install_github("xukaili/ggcor",upgrade = F)
datp <- cor.mtest(M)$p[1:48,49:51] %>%
  as.data.frame()%>% 
  rownames_to_column("A") %>% 
  gather("B","p",-A)
data <- cor(M)[1:48,49:51] %>%
  as.data.frame()%>% 
  rownames_to_column("A") %>% 
  gather("B","r",-A) %>% 
  mutate(p = datp$p)
data$ps = case_when(data$p<0.01~"**",
                    data$p<0.05~"*",
                    T~"")
library(vctrs)
library(grid)

geom_rectriangle <- function(mapping = NULL, data = NULL,
                             stat = "identity", position = "identity",
                             ...,
                             linejoin = "mitre",
                             na.rm = FALSE,
                             show.legend = NA,
                             inherit.aes = TRUE) {
  layer(
    data = data,
    mapping = mapping,
    stat = stat,
    geom = GeomRectriangle,
    position = position,
    show.legend = show.legend,
    inherit.aes = inherit.aes,
    params = list(
      linejoin = linejoin,
      na.rm = na.rm,
      ...
    )
  )
}

GeomRectriangle <- ggproto(
  "GeomRectriangle", Geom,
  default_aes = aes(r = 1, colour = "grey35", fill = NA, size = 0.25, linetype = 1,
                    alpha = NA,type = "upper"),
  required_aes = c("x", "y"),
  draw_panel = function(self, data, panel_params, coord, linejoin = "mitre",type = "upper") {
    aesthetics <- setdiff(names(data), c("x", "y"))
    
    polys <- lapply(split(data, seq_len(nrow(data))), function(row) {
      rectriangle <- point_to_rectriangle(row$x, row$y, row$r, row$type)
      aes <- new_data_frame(row[aesthetics])[rep(1, 4), ]
      GeomPolygon$draw_panel(cbind(rectriangle, aes), panel_params, coord)
    })
    
    ggplot2:::ggname("geom_rectriangle", do.call("grobTree", polys))
  },
  draw_key = draw_key_polygon
)


point_to_rectriangle <- function(x, y, r, type = type) {
  r <- 0.5 * sign(r) * sqrt(abs(r))
  #r0 = 0.5
  xmin <- - r + x
  xmax <- r + x
  ymin <- - r + y
  ymax <- r + y 
  if(type == "upper"){
    df = new_data_frame(list(
      y = c(ymax, ymax, ymin, ymax),
      x = c(xmin, xmax, xmin, xmin)
    ))
  }else if(type == "lower"){
    df = new_data_frame(list(
      y = c(ymax, ymin, ymin, ymax),
      x = c(xmax, xmax, xmin, xmax) 
    ))
  }
  df
}

data$B=factor(data$B,levels = c('E2F6','TOP2A','PredictScore'))
table(data$B)
ggplot()+
  geom_rectriangle(data = data, aes(A, B, fill = -log10(p)),type = "upper", r = 1)+
  scale_fill_gradient(low = "white", high = "#a5d3ed",na.value = "white")+
  ggnewscale::new_scale_fill()+
  geom_rectriangle(data = data, aes(A, B, fill = r),type = "lower", r = 1)+
  scale_fill_gradient2(high = "#e7857b", mid = "white",low = "#6d9cc5")+
  labs(x = "", y = "")+
  geom_text(data = data,aes(A,B,label = ps),
            nudge_y = 0.05)+
  theme_bw()+
  theme(axis.text.x = element_text(angle = 80, hjust = 1, vjust = 1))
ggsave('figure08_zhengligeneset/icgc_ssgsea_correlation_hubgene_predict_pathway.pdf',height = 6,width =12,dpi = 320,units = 'in')

###########2.3计算icgc表型estimate评分gsva####
rm(list = ls())
library(GSVA)
load('figure08_zhengligeneset/phenotype_genesets.Rdata')
load('figure02_E2F6/icgc_lihc_sur_data.Rdata')
exprSet[1:4,1:4]
load('figure06_two_group/icgc_predict_factor_clinicalinformation.Rdata')
identical(colnames(exprSet),rownames(meta))
exp=exprSet
exp[1:4,1:4]

geneset <- df
class(exp)
exp=as.matrix(exp)####这个一定要是矩阵，不能是数据框
f = "figure08_zhengligeneset/icgc_phenotype_gsva.Rdata"
if(!file.exists(f)){
  re <- gsva(exp, geneset, method="gsva",
             mx.diff=FALSE, verbose=FALSE)
  save(re,file = f)
}
load(f)
library(pheatmap)
Cluster1_name=rownames(meta)[meta$Cluster=='Cluster1']
Cluster2_name=rownames(meta)[meta$Cluster=='Cluster2']
table(meta$Group)
Group1_name=rownames(meta)[meta$Group=='Others']
Group2_name=rownames(meta)[meta$Group=='E2F6_H&TOP2A_H']
re1=re[,Cluster1_name]
re2=re[,Cluster2_name]
re=cbind(re1,re2)
meta=meta[match(colnames(re),rownames(meta)),]
identical(rownames(meta),colnames(re))
re2=as.data.frame(re)
an = data.frame(group = meta$Cluster,
                row.names = colnames(re2))
dev.off()
re2=as.matrix(re2)
bk <- c(seq(-2,-0.1,by=0.01),seq(0,2,by=0.01))
pdf('figure08_zhengligeneset/icgc_gsva_phenotype_heatmap_cluster.pdf',height = 5,width = 8)
pheatmap(re2,scale = "row",
         show_colnames = F,
         annotation_col = an,
         cluster_cols = F,
         cluster_rows = F,column_split=an,
         #legend_breaks=seq(-2,2,1),
         breaks=bk,
         color = colorRampPalette(c("navy", "white", "firebrick3"))(100))

dev.off()

identical(colnames(re),rownames(meta))
Group=meta$Cluster
draw_boxplot(re,Group,xlab = ' ',ylab = 'Score',sort = F)
ggsave('figure08_zhengligeneset/icgc_gsva_phenotype_boxplot_cluster.pdf',height = 6,width = 8,units = 'in')

Group1_name=rownames(meta)[meta$Group=='Others']
Group2_name=rownames(meta)[meta$Group=='E2F6_H&TOP2A_H']
re1=re[,Group1_name]
re2=re[,Group2_name]
re=cbind(re1,re2)
meta=meta[match(colnames(re),rownames(meta)),]
identical(rownames(meta),colnames(re))
re2=as.data.frame(re)
an = data.frame(group = meta$Group,
                row.names = colnames(re2))
dev.off()
re2=as.matrix(re2)
bk <- c(seq(-2,-0.1,by=0.01),seq(0,2,by=0.01))
pdf('figure08_zhengligeneset/icgc_gsva_phenotype_heatmap.pdf',height = 5,width = 9)
pheatmap(re2,scale = "row",
         show_colnames = F,
         annotation_col = an,
         cluster_cols = F,
         cluster_rows = F,column_split=an,
         #legend_breaks=seq(-2,2,1),
         breaks=bk,
         color = colorRampPalette(c("navy", "white", "firebrick3"))(100))

dev.off()

identical(colnames(re),rownames(meta))
Group=meta$Group
draw_boxplot(re,Group,xlab = ' ',ylab = 'Score',sort = F)
ggsave('figure08_zhengligeneset/icgc_gsva_phenotype_boxplot.pdf',height = 6,width = 8,units = 'in')

identical(rownames(meta),colnames(re))
exp=exp[,match(colnames(re),colnames(exp))]
identical(colnames(re),colnames(exp))
exp=exp[c('E2F6','TOP2A'),]
identical(colnames(re),colnames(exp))
re=rbind(re,exp)
identical(colnames(exp),rownames(meta))
identical(colnames(re),rownames(meta))
data=as.data.frame(t(re))
data$PredictScore=meta$Predict_score
M=data

library(tidyverse)
library(corrplot)
if(!require(ggcor))devtools::install_github("xukaili/ggcor",upgrade = F)
datp <- cor.mtest(M)$p[1:29,30:32] %>%
  as.data.frame()%>% 
  rownames_to_column("A") %>% 
  gather("B","p",-A)
data <- cor(M)[1:29,30:32] %>%
  as.data.frame()%>% 
  rownames_to_column("A") %>% 
  gather("B","r",-A) %>% 
  mutate(p = datp$p)
data$ps = case_when(data$p<0.01~"**",
                    data$p<0.05~"*",
                    T~"")
library(vctrs)
library(grid)

geom_rectriangle <- function(mapping = NULL, data = NULL,
                             stat = "identity", position = "identity",
                             ...,
                             linejoin = "mitre",
                             na.rm = FALSE,
                             show.legend = NA,
                             inherit.aes = TRUE) {
  layer(
    data = data,
    mapping = mapping,
    stat = stat,
    geom = GeomRectriangle,
    position = position,
    show.legend = show.legend,
    inherit.aes = inherit.aes,
    params = list(
      linejoin = linejoin,
      na.rm = na.rm,
      ...
    )
  )
}

GeomRectriangle <- ggproto(
  "GeomRectriangle", Geom,
  default_aes = aes(r = 1, colour = "grey35", fill = NA, size = 0.25, linetype = 1,
                    alpha = NA,type = "upper"),
  required_aes = c("x", "y"),
  draw_panel = function(self, data, panel_params, coord, linejoin = "mitre",type = "upper") {
    aesthetics <- setdiff(names(data), c("x", "y"))
    
    polys <- lapply(split(data, seq_len(nrow(data))), function(row) {
      rectriangle <- point_to_rectriangle(row$x, row$y, row$r, row$type)
      aes <- new_data_frame(row[aesthetics])[rep(1, 4), ]
      GeomPolygon$draw_panel(cbind(rectriangle, aes), panel_params, coord)
    })
    
    ggplot2:::ggname("geom_rectriangle", do.call("grobTree", polys))
  },
  draw_key = draw_key_polygon
)


point_to_rectriangle <- function(x, y, r, type = type) {
  r <- 0.5 * sign(r) * sqrt(abs(r))
  #r0 = 0.5
  xmin <- - r + x
  xmax <- r + x
  ymin <- - r + y
  ymax <- r + y 
  if(type == "upper"){
    df = new_data_frame(list(
      y = c(ymax, ymax, ymin, ymax),
      x = c(xmin, xmax, xmin, xmin)
    ))
  }else if(type == "lower"){
    df = new_data_frame(list(
      y = c(ymax, ymin, ymin, ymax),
      x = c(xmax, xmax, xmin, xmax) 
    ))
  }
  df
}

data$B=factor(data$B,levels = c('E2F6','TOP2A','PredictScore'))
table(data$B)
ggplot()+
  geom_rectriangle(data = data, aes(A, B, fill = -log10(p)),type = "upper", r = 1)+
  scale_fill_gradient(low = "white", high = "#a5d3ed",na.value = "white")+
  ggnewscale::new_scale_fill()+
  geom_rectriangle(data = data, aes(A, B, fill = r),type = "lower", r = 1)+
  scale_fill_gradient2(high = "#e7857b", mid = "white",low = "#6d9cc5")+
  labs(x = "", y = "")+
  geom_text(data = data,aes(A,B,label = ps),
            nudge_y = 0.05)+
  theme_bw()+
  theme(axis.text.x = element_text(angle = 80, hjust = 1, vjust = 1))
ggsave('figure08_zhengligeneset/icgc_gsva_correlation_hubgene_predict_phenotype.pdf',height = 4,width =8,dpi = 320,units = 'in')

##########2.4计算tcga通路estimate评分gsva####
rm(list = ls())
library(GSVA)
load('figure08_zhengligeneset/pathways_genesets.Rdata')
load('figure02_E2F6/icgc_lihc_sur_data.Rdata')
exprSet[1:4,1:4]
load('figure06_two_group/icgc_predict_factor_clinicalinformation.Rdata')
identical(colnames(exprSet),rownames(meta))
exp=exprSet
exp[1:4,1:4]

geneset <- pathways
class(exp)
exp=as.matrix(exp)####这个一定要是矩阵，不能是数据框
f = "figure08_zhengligeneset/icgc_pathway_gsva.Rdata"
if(!file.exists(f)){
  re <- gsva(exp, geneset, method="gsva",
             mx.diff=FALSE, verbose=FALSE)
  save(re,file = f)
}
load(f)
library(pheatmap)
Cluster1_name=rownames(meta)[meta$Cluster=='Cluster1']
Cluster2_name=rownames(meta)[meta$Cluster=='Cluster2']
table(meta$Group)
Group1_name=rownames(meta)[meta$Group=='Others']
Group2_name=rownames(meta)[meta$Group=='E2F6_H&TOP2A_H']
re1=re[,Cluster1_name]
re2=re[,Cluster2_name]
re=cbind(re1,re2)
meta=meta[match(colnames(re),rownames(meta)),]
identical(rownames(meta),colnames(re))
re2=as.data.frame(re)
an = data.frame(group = meta$Cluster,
                row.names = colnames(re2))
dev.off()
re2=as.matrix(re2)
bk <- c(seq(-2,-0.1,by=0.01),seq(0,2,by=0.01))
pdf('figure08_zhengligeneset/icgc_gsva_pathway_heatmap_cluster.pdf',height = 8,width = 8)
pheatmap(re2,scale = "row",
         show_colnames = F,
         annotation_col = an,
         cluster_cols = F,
         cluster_rows = F,column_split=an,
         #legend_breaks=seq(-2,2,1),
         breaks=bk,
         color = colorRampPalette(c("navy", "white", "firebrick3"))(100))

dev.off()

identical(colnames(re),rownames(meta))
Group=meta$Cluster
draw_boxplot(re,Group,xlab = ' ',ylab = 'Score',sort = F)
ggsave('figure08_zhengligeneset/icgc_gsva_pathway_boxplot_cluster.pdf',height = 9,width = 10,units = 'in')

Group1_name=rownames(meta)[meta$Group=='Others']
Group2_name=rownames(meta)[meta$Group=='E2F6_H&TOP2A_H']
re1=re[,Group1_name]
re2=re[,Group2_name]
re=cbind(re1,re2)
meta=meta[match(colnames(re),rownames(meta)),]
identical(rownames(meta),colnames(re))
re2=as.data.frame(re)
an = data.frame(group = meta$Group,
                row.names = colnames(re2))
dev.off()
re2=as.matrix(re2)
bk <- c(seq(-2,-0.1,by=0.01),seq(0,2,by=0.01))
pdf('figure08_zhengligeneset/icgc_gsva_pathway_heatmap.pdf',height = 8,width = 10)
pheatmap(re2,scale = "row",
         show_colnames = F,
         annotation_col = an,
         cluster_cols = F,
         cluster_rows = F,column_split=an,
         #legend_breaks=seq(-2,2,1),
         breaks=bk,
         color = colorRampPalette(c("navy", "white", "firebrick3"))(100))

dev.off()

identical(colnames(re),rownames(meta))
Group=meta$Group
draw_boxplot(re,Group,xlab = ' ',ylab = 'Score',sort = F)
ggsave('figure08_zhengligeneset/icgc_gsva_pathway_boxplot.pdf',height = 9,width = 12,units = 'in')

identical(rownames(meta),colnames(re))
exp=exp[,match(colnames(re),colnames(exp))]
identical(colnames(re),colnames(exp))
exp=exp[c('E2F6','TOP2A'),]
identical(colnames(re),colnames(exp))
re=rbind(re,exp)
identical(colnames(exp),rownames(meta))
identical(colnames(re),rownames(meta))
data=as.data.frame(t(re))
data$PredictScore=meta$Predict_score
M=data

library(tidyverse)
library(corrplot)
if(!require(ggcor))devtools::install_github("xukaili/ggcor",upgrade = F)
datp <- cor.mtest(M)$p[1:48,49:51] %>%
  as.data.frame()%>% 
  rownames_to_column("A") %>% 
  gather("B","p",-A)
data <- cor(M)[1:48,49:51] %>%
  as.data.frame()%>% 
  rownames_to_column("A") %>% 
  gather("B","r",-A) %>% 
  mutate(p = datp$p)
data$ps = case_when(data$p<0.01~"**",
                    data$p<0.05~"*",
                    T~"")
library(vctrs)
library(grid)

geom_rectriangle <- function(mapping = NULL, data = NULL,
                             stat = "identity", position = "identity",
                             ...,
                             linejoin = "mitre",
                             na.rm = FALSE,
                             show.legend = NA,
                             inherit.aes = TRUE) {
  layer(
    data = data,
    mapping = mapping,
    stat = stat,
    geom = GeomRectriangle,
    position = position,
    show.legend = show.legend,
    inherit.aes = inherit.aes,
    params = list(
      linejoin = linejoin,
      na.rm = na.rm,
      ...
    )
  )
}

GeomRectriangle <- ggproto(
  "GeomRectriangle", Geom,
  default_aes = aes(r = 1, colour = "grey35", fill = NA, size = 0.25, linetype = 1,
                    alpha = NA,type = "upper"),
  required_aes = c("x", "y"),
  draw_panel = function(self, data, panel_params, coord, linejoin = "mitre",type = "upper") {
    aesthetics <- setdiff(names(data), c("x", "y"))
    
    polys <- lapply(split(data, seq_len(nrow(data))), function(row) {
      rectriangle <- point_to_rectriangle(row$x, row$y, row$r, row$type)
      aes <- new_data_frame(row[aesthetics])[rep(1, 4), ]
      GeomPolygon$draw_panel(cbind(rectriangle, aes), panel_params, coord)
    })
    
    ggplot2:::ggname("geom_rectriangle", do.call("grobTree", polys))
  },
  draw_key = draw_key_polygon
)


point_to_rectriangle <- function(x, y, r, type = type) {
  r <- 0.5 * sign(r) * sqrt(abs(r))
  #r0 = 0.5
  xmin <- - r + x
  xmax <- r + x
  ymin <- - r + y
  ymax <- r + y 
  if(type == "upper"){
    df = new_data_frame(list(
      y = c(ymax, ymax, ymin, ymax),
      x = c(xmin, xmax, xmin, xmin)
    ))
  }else if(type == "lower"){
    df = new_data_frame(list(
      y = c(ymax, ymin, ymin, ymax),
      x = c(xmax, xmax, xmin, xmax) 
    ))
  }
  df
}

data$B=factor(data$B,levels = c('E2F6','TOP2A','PredictScore'))
table(data$B)
ggplot()+
  geom_rectriangle(data = data, aes(A, B, fill = -log10(p)),type = "upper", r = 1)+
  scale_fill_gradient(low = "white", high = "#a5d3ed",na.value = "white")+
  ggnewscale::new_scale_fill()+
  geom_rectriangle(data = data, aes(A, B, fill = r),type = "lower", r = 1)+
  scale_fill_gradient2(high = "#e7857b", mid = "white",low = "#6d9cc5")+
  labs(x = "", y = "")+
  geom_text(data = data,aes(A,B,label = ps),
            nudge_y = 0.05)+
  theme_bw()+
  theme(axis.text.x = element_text(angle = 45, hjust = 1, vjust = 1))
ggsave('figure08_zhengligeneset/icgc_gsva_correlation_hubgene_predict_pathway.pdf',height = 6,width =12,dpi = 320,units = 'in')
